% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancePMV7730.r
\name{balancePMV7730}
\alias{balancePMV7730}
\title{balancePMV7730}
\usage{
balancePMV7730(pmv, M)
}
\arguments{
\item{numeric}{pmv Fanger ISO 7730 Predicted Mean Vote.}

\item{numeric}{M Metabolism of the person.}
}
\value{

}
\description{
Calculate the thermal balance of the body in W/m.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{balancePMV7730}

