% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSA.r
\name{BSA}
\alias{BSA}
\title{BSA}
\usage{
BSA(h, w)
}
\arguments{
\item{h}{numeric Body height in cm.}

\item{w}{numeric Body weight in kg.}
}
\value{
BSA index
}
\description{
Calculate the Body Surface Area by using anthropometric features (height and weight).
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{Area}
\keyword{Body}
\keyword{Surface}
