% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2F.r
\name{C2F}
\alias{C2F}
\title{C2F}
\usage{
C2F(C)
}
\arguments{
\item{C}{numeric Temperature in degC.}
}
\value{

}
\description{
Convert temperature from Celsius degree (degC) to Fahrenheit degree (degF).
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{temperature}
