% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clomax_7730.r
\name{clomax_7730}
\alias{clomax_7730}
\title{clomax_7730}
\usage{
clomax_7730(t, rh, wind, tr, M, W = 0)
}
\arguments{
\item{t}{numeric    Air temperature in degC.}

\item{rh}{numeric   Air Relative humidity in percentage.}

\item{wind}{numeric Wind speed in meters per second.}

\item{tr}{numeric   Mean radiant temperature in degC.}

\item{M}{numeric    Metabolic rate of subject in W/mq.}

\item{W}{numeric    Mechanical work of subject in W/mq. Generally 0.}
}
\value{
maximum  clothing insulation
}
\description{
Calculate maximum clothing insulation value required to reach thermal comfort in moderate thermal environments ( -0.5 < PMV <0.5) based on PMV ISO 7730.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
