% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clomin_custom.r
\name{clomin_custom}
\alias{clomin_custom}
\title{clomin_custom}
\usage{
clomin_custom(t, rh, wind, tr, age, mbody, ht, gender = "male")
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Windspeed in meter per second.}

\item{numeric}{tr Air temperature in Celsius degrees.}

\item{numeric}{M Metabolic rate of subject Watt per mq.}

\item{numeric}{W Work rate of subject in Watt per mq.}

\item{numeric}{age Age in years.}

\item{numeric}{mbody Body Mass in kg.}

\item{numeric}{ht Heigth of subject in meters.}

\item{character}{gender Sex of the subject. Default is "male" otherwise "female".}
}
\value{

}
\description{
Calculate minimal clothing insulation value needed for thermal comfort.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{ocir}
