% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_16.r
\name{compass_16}
\alias{compass_16}
\title{compass_16}
\usage{
compass_16(direction)
}
\arguments{
\item{numeric}{direction Directions in degrees from the North ( meteorological) .}
}
\value{

}
\description{
Calculate thewind sector (16 partition)  of its provenience.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{compass_16}

