% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_8.r
\name{compass_8}
\alias{compass_8}
\title{compass_8}
\usage{
compass_8(direction)
}
\arguments{
\item{direction}{numeric   Directions in degrees from the North.}
}
\value{
sector character
}
\description{
Calculate the  wind sector (8 partition)  provenience.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
