% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deficitsat.r
\name{deficitsat}
\alias{deficitsat}
\title{deficitsat}
\usage{
deficitsat(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{

}
\description{
Computes saturation deficit in hPa. This is air dryness evaluation.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
