% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degtorad.r
\name{degtorad}
\alias{degtorad}
\title{degtorad}
\usage{
degtorad(angledeg)
}
\arguments{
\item{angledeg}{numeric Angle in decimal degrees.}
}
\value{

}
\description{
Conversion angles from decimal degrees to radians.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{degtorad}
