% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewpoint.r
\name{dewpoint}
\alias{dewpoint}
\title{dewpoint}
\usage{
dewpoint(t, rh, formula = "NOAA")
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Air Relative humidity in percentage.}

\item{formula}{character  Default is "NOAA".}
}
\value{

}
\description{
Computes the dewpoint temperature in degC   following different computational schemes defined by formula of saturation pressure estimation
("NOAA","Sonntag","Paroscientific").
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
