% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusivity.r
\name{diffusivity}
\alias{diffusivity}
\title{diffusivity}
\usage{
diffusivity(tk, pair)
}
\arguments{
\item{tk}{numeric Air temperature in degK.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{

}
\description{
Calculate air diffusivity.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{diffusivity}
