% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discomfort_index.r
\name{discomfort_index}
\alias{discomfort_index}
\title{discomfort_index}
\usage{
discomfort_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Given t air temperature (Celsius), rh relative humidity and gives  Discomfort index (DI). It belongs to direct indexes class.
}
\references{
Thom, E.C.,1959,The discomfort index. Weather wise, 12, 57–60.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{discomfort_index}
