% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enthalpy.r
\name{enthalpy}
\alias{enthalpy}
\title{enthalpy}
\usage{
enthalpy(t, hum_ratio)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{hum_ratio}{numeric Humidity ratio.}
}
\value{

}
\description{
Calculates the enthalpy of air/water vapor mixture [kJ/kg]. Inputs are dry-bulb temperature [degC] and humidity ratio [unitless].
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{enthalpy}
