% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.r
\name{es}
\alias{es}
\title{es}
\usage{
es(t)
}
\arguments{
\item{t}{numeric Air temperature in Celsius degrees.}
}
\value{
Saturation Vapor Pressure in hPa.
}
\description{
Given a air temperature  in degC gives Vapor Pressure in hPa. Hardy, R.; ITS-90 Formulations for Vapor Pressure, Frostpoint
Temperature, Dewpoint Temperature and Enhancement Factors in the Range -100 to 100 degC;
}
\references{
Proceedings of Third International Symposium on Humidity and Moisture. National Physical Laboratory (NPL), London, 1998, pp. 214-221 \url{http://www.thunderscientific.com/tech_info/reflibrary/its90formulas.pdf}
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{air}
\keyword{saturation}
