% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2C.r
\name{F2C}
\alias{F2C}
\title{F2C}
\usage{
F2C(F)
}
\arguments{
\item{F}{numeric Temperature in Fahrenheit degrees (degF)}
}
\value{
Celsius degrees
}
\description{
Converts temperature from Fahrenheit degree (degF)  to Celsius degree (degC).
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{F2C}
