% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits_index.r
\name{fits_index}
\alias{fits_index}
\title{fits_index}
\usage{
fits_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Calculate Fighter Index of Thermal Stress
}
\references{
Nunneley SH, Stribley F,1979, Fighter index of thermal stress FITS: guidance for hot-weather aircraft operations.Aviat Space Environ Med 50, 639–42.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{,}
\keyword{Fighter}
\keyword{Index}
\keyword{Stress}
\keyword{Thermal}
\keyword{fits_index}
\keyword{of}
