% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRT.r
\name{fMRT}
\alias{fMRT}
\title{fMRT}
\usage{
fMRT(t, solar)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rad Global short solar irradiance in W/mq.}

\item{numeric}{frad Fraction related to direct solar beam [0-1].}
}
\value{
Mean Radiant temperature.
}
\description{
Calculated the mean radiant temperature from the solar radiation.  Modified based on direct && diffuse ratio. Assumes a uniform surround temperature of Ta && short wave solar radiation only;
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{fMRT}

