% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRH.r
\name{fRH}
\alias{fRH}
\title{fRH}
\usage{
fRH(tair, dewp)
}
\arguments{
\item{numeric}{tair Vector or value of air temperature in Celsius degree.}

\item{numeric}{dewp Vector or value of dew point temperature in Celsius degree.}
}
\value{
numeric  - centroid of the input features
}
\description{
Return relative humidity from air temperature (Celsius) and Dew point (Celsius);
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Relative}
\keyword{humidity}

