% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frostime.r
\name{frostime}
\alias{frostime}
\title{frostime}
\usage{
frostime(t, wind)
}
\arguments{
\item{t}{numeric Air temperature in Celsius degrees.}

\item{wind}{numeric the wind speed in meter per seconds [m/s].}
}
\value{
time of frosting
}
\description{
Given a temperature t (Celsius) and wind ( m/sec) frost time following Wind Chill Frostbite Chart.
}
\references{
ISO 9920:2007 Ergonomics of the thermal environment -- Estimation of thermal insulation and water vapour resistance of a clothing ensemble.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{frostime}
