% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTmrt.r
\name{fTmrt}
\alias{fTmrt}
\title{fTmrt}
\usage{
fTmrt(t, Tg, wind)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{Tg Globe temperature in Celsius degrees.}

\item{numeric}{wind Speed or wind velocity in meter per secons [m/s].}
}
\value{

}
\description{
Calculate the mean radiant temperature by using Thorrson scheme.
}
\references{
Different methods for estimating the mean radiant temperature in an outdoor urban setting \url{http://onlinelibrary.wiley.com/doi/10.1002/joc.1537/full}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{fTmrt}
