% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts2knots.r
\name{fts2knots}
\alias{fts2knots}
\title{fts2knots}
\usage{
fts2knots(fts)
}
\arguments{
\item{fts}{numeric  Speed in feets per second.}
}
\value{
knots per second
}
\description{
Conversion from feet per second to knots per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{fts2knots}
