% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts2mph.r
\name{fts2mph}
\alias{fts2mph}
\title{fts2mph}
\usage{
fts2mph(fts)
}
\arguments{
\item{fts}{numeric  Speed in feets per second.}
}
\value{
miles per hour
}
\description{
Conversion from feet per second to miles per hour.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{fts2mph}
