% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts2ms.r
\name{fts2ms}
\alias{fts2ms}
\title{fts2ms}
\usage{
fts2ms(fts)
}
\arguments{
\item{fts}{numeric  Speed in feets per second.}
}
\value{

}
\description{
Conversion from feet per second to meter per second.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{fts2ms}
