% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDOY.r
\name{getDOY}
\alias{getDOY}
\title{getDOY}
\usage{
getDOY(date)
}
\arguments{
\item{date}{Date in format YYYY-MM-DD.}
}
\value{
numeric Julian day.
}
\description{
Get julian day number of date.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{getDOY}
