% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_risk_hexrgb_level.r
\name{heat_risk_hexrgb_level}
\alias{heat_risk_hexrgb_level}
\title{heat_risk_hexrgb_level}
\usage{
heat_risk_hexrgb_level(wbgt, cav, tresh)
}
\arguments{
\item{wbgt}{numeric Wetbulb globe temperature index in degC}

\item{cav}{numeric Clothing adjusted value due to worker clothing ensemble}

\item{tresh}{numeric treshshold for heat risk in degC ( RAL or REL)}
}
\value{

}
\description{
Calculate the heat risk level for worker by using color codes returning hex code.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{}
