% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_risk_level.r
\name{heat_risk_level}
\alias{heat_risk_level}
\title{heat_risk_level}
\usage{
heat_risk_level(wbgt, cav, tresh)
}
\arguments{
\item{wbgt}{numeric Wetbulb globe temperature index in degC}

\item{cav}{numeric Clothing adjusted value due to worker clothing ensemble}

\item{tresh}{numeric treshshold for heat risk in degC ( RAL or REL)}
}
\value{
numeric level risk
}
\description{
Calculate the heat risk level for worker as four classes level.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
