% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_risk_text_level_eng.r
\name{heat_risk_text_level_eng}
\alias{heat_risk_text_level_eng}
\title{heat_risk_text_level_eng}
\usage{
heat_risk_text_level_eng(wbgt, cav, tresh)
}
\arguments{
\item{wbgt}{numeric Wetbulb globe temperature index in degC}

\item{cav}{numeric Clothing adjusted value due to worker clothing ensemble}

\item{tresh}{numeric treshshold for heat risk in degC ( RAL or REL)}
}
\value{
level risk by color name in ENG.
}
\description{
Calculate the heat risk level for worker by using color codes.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
