% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_index.r
\name{hi_index}
\alias{hi_index}
\title{hi_index}
\usage{
hi_index(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage(%).}
}
\value{

}
\description{
Given air temperature (Celsius), relative humidity  gives a heat stress index in Celsius degree for primates animals.
}
\references{
Hill et al,2004, Indices of environmental temperatures for primates in open habitats.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{hi_index}
