% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSI_index.r
\name{HSI_index}
\alias{HSI_index}
\title{HSI_index}
\usage{
HSI_index(t, rh, wind, tr, M, W, clo, param = "HSI")
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity.}

\item{numeric}{wind Windspeed in meter per second.}

\item{numeric}{tr Mean radiant temperature in Celsius degrees.}

\item{numeric}{M Metabolic rate of subject Watt on mq.}

\item{numeric}{W Work rate of subject Watt on mq.}

\item{numeric}{clo Clothing insulation level in clo.}

\item{character}{param index requested HSI or AET Allowable exposure time in minutes.}
}
\value{

}
\description{
Computes the Heat Strain Index. HSI is the ratio of the demand for sweat evaporation to capacity of evaporation (Ereq on Emax). This denotes also the percent of skin wettedness, which is a good predictor of warm discomfort.
}
\references{
Belding HS, Hatch TF, 1955. Index for evaluating heat stress in terms of resulting physiological strain. Heat Pip Air Condit 27, 129–36.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{HSI}
