% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidex.r
\name{humidex}
\alias{humidex}
\title{humidex}
\usage{
humidex(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Computes canadian Humidex index.
}
\references{
Masterson, J. et Richardson, F. A., Humidex, A Method of Quantifying Human Discomfort Due to Excessive Heat and Humidity, Downsview, Ontario, Environnement Canada, 1979, 45.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{canada}
\keyword{humidex,}
