% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humrat.r
\name{humrat}
\alias{humrat}
\title{humrat}
\usage{
humrat(p, pair)
}
\arguments{
\item{p}{numeric  Water pressure in hPa.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{
humidity ratio
}
\description{
Calculates humidity ratio in vapor-air mixture.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{humrat}
