% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLeapYear.r
\name{isLeapYear}
\alias{isLeapYear}
\title{isLeapYear}
\usage{
isLeapYear(yr)
}
\arguments{
\item{yr}{numeric Year.}
}
\value{
boolean [0 - 1] indicate if year is leap
}
\description{
Logical if the year indicated is a leap year.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{isLeapYear}
