% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2fts.r
\name{kmh2fts}
\alias{kmh2fts}
\title{kmh2fts}
\usage{
kmh2fts(kmh)
}
\arguments{
\item{kmh}{numeric  Speed in kilometer per hour.}
}
\value{
feets per second
}
\description{
Conversion from kilometer per hour to feets per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
