% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2knots.r
\name{kmh2knots}
\alias{kmh2knots}
\title{kmh2knots}
\usage{
kmh2knots(kmh)
}
\arguments{
\item{kmh}{numeric  Speed in kilometer per hour.}
}
\value{

}
\description{
Conversion from kilometer per hour to knot per second.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{kmh2knots}
