% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2mph.r
\name{kmh2mph}
\alias{kmh2mph}
\title{kmh2mph}
\usage{
kmh2mph(kmh)
}
\arguments{
\item{kmh}{numeric  Speed in Kilometer per hours.}
}
\value{
miles per hour
}
\description{
Conversion kilometer per hour to miles per hour.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{kmh2mph}
