% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2ms.r
\name{kmh2ms}
\alias{kmh2ms}
\title{kmh2ms}
\usage{
kmh2ms(kmh)
}
\arguments{
\item{kmh}{numeric  Speed kilometer per hour.}
}
\value{
meters per second
}
\description{
Conversion from kilometers per hour to meters per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{kmh2ms}
