% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2kmh.r
\name{knots2kmh}
\alias{knots2kmh}
\title{knots2kmh}
\usage{
knots2kmh(knots)
}
\arguments{
\item{numeric}{knots Speed in knot per second.}
}
\value{

}
\description{
Conversion from knot per second to kilometer per hour.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{knots2kmh}
