% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lost_productivity.r
\name{lost_productivity}
\alias{lost_productivity}
\title{lost_productivity}
\usage{
lost_productivity(wbgt, tresh)
}
\arguments{
\item{wbgt}{numeric Wetbulb globe temperature index in degC}

\item{tresh}{numeric treshshold for loss in degC}
}
\value{
percentage of productivity lost
}
\description{
Calculate the percentage in worker's lost productivity due to heat conditions.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{Area}
\keyword{Body}
\keyword{Surface}
