% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtoinHG.r
\name{mbtoinHG}
\alias{mbtoinHG}
\title{mbtoinHG}
\usage{
mbtoinHG(mb)
}
\arguments{
\item{mb}{numeric   Air pressure in Millibar [hPa]}
}
\value{

}
\description{
Conversion from Millibar [hPa] to inches HG.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{mbtoinHG}
