% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtolbsinch.r
\name{mbtolbsinch}
\alias{mbtolbsinch}
\title{mbtolbsinch}
\usage{
mbtolbsinch(mb)
}
\arguments{
\item{mb}{numeric   Air pressure in Millibar [hPa].}
}
\value{
inches
}
\description{
Conversion from Millibar [hPa] to inches.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{mbtolbsinch}
