% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtommHG.r
\name{mbtommHG}
\alias{mbtommHG}
\title{mbtommHG}
\usage{
mbtommHG(mb)
}
\arguments{
\item{mb}{numeric   Air pressure in Millibar [hPa].}
}
\value{

}
\description{
Conversion from Millibar [hPa] to mmHG.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{mbtommHG}
