% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdi_index.r
\name{mdi_index}
\alias{mdi_index}
\title{mdi_index}
\usage{
mdi_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{
mdi_index
}
\description{
Modified discomfort index MDI
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Moran DS, Shitzer A, Pandolf KB (1998) A physiological strain index to evaluate heat stress. Am J Physiol 275, R129–34
}
\keyword{mdi_index}

