% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_rate.r
\name{met_rate}
\alias{met_rate}
\title{met_rate}
\usage{
met_rate(BSA, isolevel)
}
\arguments{
\item{BSA}{numeric Body are surface.}

\item{isolevel}{numeric Metabolic rate class following level of ISO FDIS 7243.}
}
\value{
basal metabolic rate
}
\description{
Calculate metabolic
}
\references{
Ergonomics of the thermal environment – Assessment of heat stress using the WBGT (wet bulb globe temperature)   index, ISO FDIS 7243 (2016)
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{MET}
