% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2kmh.r
\name{mph2kmh}
\alias{mph2kmh}
\title{mph2kmh}
\usage{
mph2kmh(mph)
}
\arguments{
\item{numeric}{mph Speed Miles on hours.}
}
\value{

}
\description{
Conversion speed in  miles per hour to kilometers per hour.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{mph2kmh}
