% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2knots.r
\name{mph2knots}
\alias{mph2knots}
\title{mph2knots}
\usage{
mph2knots(mph)
}
\arguments{
\item{mph}{numeric  Speed miles per hour.}
}
\value{
knots per second
}
\description{
Conversion speed in  miles per hour to knots per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{mph2knots}
