% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_bernard}
\alias{mrt_bernard}
\title{mrt_bernard}
\usage{
mrt_bernard(t, tg, wind, diam)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{tg Global short solar irradiance in W on mq.}

\item{numeric}{wind Fraction related to direct solar beam 0-1.}
}
\value{
Mean Radiant temperature.
}
\description{
Calculated the mean radiant temperature from the solar radiation by using the bernard scheme.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{MRT}
