% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_globe}
\alias{mrt_globe}
\title{mrt_globe}
\usage{
mrt_globe(t, tg, wind, diam)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{tg Global short solar irradiance in Watt on mq.}

\item{numeric}{wind Windspeed in meter per second.}

\item{numeric}{diam diameter of the sphere in millimeter. Input example  50 mm.}
}
\value{
Mean Radiant temperature.
}
\description{
Given air temperature , the globe Temeperature, the wind speed and diameter mean radiant temperature is done according ISO 7726 1998.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{globe}
\keyword{temperature}
