% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_solar_proj}
\alias{mrt_solar_proj}
\title{mrt_solar_proj}
\usage{
mrt_solar_proj(t, rh, solar, sunelev, albedo, fdir, emis_sfc)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rad Global short solar irradiance in Watt on mq.}

\item{numeric}{frad Fraction related to direct solar beam 0 to 1.}
}
\value{
Mean Radiant temperature.
}
\description{
Calculated the mean radiant temperature from the short solar irradiance with human projection factor. it assumes a uniform surround temperature of air temperature  and short wave solar radiation.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{fMRT}
