% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_thorsson}
\alias{mrt_thorsson}
\title{mrt_thorsson}
\usage{
mrt_thorsson(t, tg, wind, diam)
}
\arguments{
\item{numeric}{t air temperature in Celsius degrees.}

\item{numeric}{tg global short solar irradiance in Watt on mq.}

\item{numeric}{wind windspeed in meter per second.}

\item{numeric}{diam diameter of the sphere in millimeter. Input example  50 mm.}
}
\value{
Mean Radiant temperature.
}
\description{
Calculated the mean radiant temperature from the solar radiation.  Modified based on direct and diffuse ratio. Assumes a uniform surround temperature of Ta && short wave solar radiation only;
}
\references{
Thorsson et al, Different methods for estimating the mean radiant temperature in an outdoor urban setting, Int. J. Climatol.27  1983   1993 (2007)
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{MRT}
