% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2fts.r
\name{ms2fts}
\alias{ms2fts}
\title{ms2fts}
\usage{
ms2fts(ms)
}
\arguments{
\item{ms}{Speed numeric   in meters per second [m/s].}
}
\value{

}
\description{
Conversion speed meter per second in feet per second.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{ms2fts}
