% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2kmh.r
\name{ms2kmh}
\alias{ms2kmh}
\title{ms2kmh}
\usage{
ms2kmh(ms)
}
\arguments{
\item{ms}{Speed numeric   in meters per second [m/s].}
}
\value{
kilometers per hour
}
\description{
Conversion speed meter per second  in kilometers per hour.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{ms2kmh}
