% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2knots.r
\name{ms2knots}
\alias{ms2knots}
\title{ms2knots}
\usage{
ms2knots(ms)
}
\arguments{
\item{ms}{Speed numeric   in meters per second [m/s].}
}
\value{
knots per second
}
\description{
Conversion speed meter per second in knots per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{ms2knots}
