% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_index.r
\name{net_index}
\alias{net_index}
\title{net_index}
\usage{
net_index(t, rh, wind)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Windspeed in meter per second [m/s].}
}
\value{

}
\description{
Net effective temperature (NET) incorporates ambient temperature, wind speed, and relative humidity and is used in both hot and cold situations.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Cho KLA, Margaret Loughnan M, and Tapper T,(2013), An Exploration of Temperature Metrics for Further Developing the Heat-Health Weather Warning System in Hong Kong
}
\keyword{net_index}

