% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocir_custom.r
\name{ocir_custom}
\alias{ocir_custom}
\title{ocir_custom}
\usage{
ocir_custom(t, rh, wind, mtrad, age, mbody, ht, gender = "male")
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Wind speed in meter per second.}

\item{numeric}{tr Mean radiant temperature in Celsius degrees.}

\item{numeric}{mbody  mass of the body in kg.}

\item{numeric}{age Age of the subject in year.}

\item{numeric}{ht Height of subject in meter.}

\item{character}{gender Sex of subject. "male" or "female"-}
}
\value{

}
\description{
Calculate ocir index consisting in range of clothing insulation in moderate thermal environments based on adapted PMV scheme with individuals features.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{ocir_custom}
