% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_index.r
\name{oxford_index}
\alias{oxford_index}
\title{oxford_index}
\usage{
oxford_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Computes  oxford index WD index of discomfort.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Lind AR, Hallon RF (1957) Assessment of physiologic severity of hot climate. J Appl Physiol 11, 35–40.
}
\keyword{oxford_index}

