% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_atm_z.r
\name{p_atm_z}
\alias{p_atm_z}
\title{p_atm_z}
\usage{
p_atm_z(z, p_sl = 1013.25)
}
\arguments{
\item{z}{numeric elevation/altitude (m).}

\item{p_sl}{numeric pressure at sea level, per weather report (hPa),Default 101235 Pa.}
}
\value{
numeric pressure (Pa)
}
\description{
Calculate local air ressure as function of elevation/altitude in meters and sea-level pressure.
}
\references{
2005 ASHRAE Handbook - Fundamentals (SI) ch.6 (3)
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
