% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_local.r
\name{p_local}
\alias{p_local}
\title{p_local}
\usage{
p_local(press, topo, temp)
}
\arguments{
\item{numeric}{press Air local pressure in hPa [millibar].}

\item{numeric}{topo Altitude or Elevation in meters.}

\item{numeric}{t  Mean Air temperature in Celsius degrees.}
}
\value{

}
\description{
DESCRIPTION
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{p_local}

