% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseISO8601String.r
\name{parseISO8601String}
\alias{parseISO8601String}
\title{parseISO8601String}
\usage{
parseISO8601String(dateString)
}
\arguments{
\item{date}{Date in format YYYY-MM-DD.String}
}
\value{
date
}
\description{
Parse date in ISO8601 format.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{parseISO8601String}
