% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pierceSET.r
\name{pierceSET}
\alias{pierceSET}
\title{pierceSET}
\usage{
pierceSET(t, rh, vel, tr, M, W, clo, patm)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Windspeed in meter per second.}

\item{numeric}{tr Mean radiant temperature in Celsius degrees.}

\item{numeric}{M metabolic rate in W/mq.}

\item{numeric}{W Workload in W/mq.Default is 0.}

\item{numeric}{clo Clothing level in clo value.}

\item{numeric}{patm Air pressure in hPa.}
}
\value{
pierce SET in celsius degree.
}
\description{
Calculates standard effective temperature of Pierce. The calculation procedure is taken from CBE Thermal Comfort Tool js library. 
The Original paper of idex Gagge, A.P., Fobelets, A.P., Berglund, L. G., “A Standard Predictive Index of Human Response to the Thermal Environment”, ASHRAE Trans., Vol.92, Pt 2, 1986.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Hoyt Tyler, Schiavon Stefano, Piccioli Alberto, Moon Dustin, and Steinfeld Kyle, 2013, CBE Thermal Comfort Tool.
Center for the Built Environment, University of California Berkeley, \link{http://cbe.berkeley.edu/comforttool/}. Code is licenced CC-BY-NC.
}
\keyword{Comfort}
\keyword{Thermal}
\keyword{Tool.}
\keyword{pierceSET,CBE}

