% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMV_ISO7730.r
\name{PMV_ISO7730}
\alias{PMV_ISO7730}
\title{PMV_ISO7730}
\usage{
PMV_ISO7730(t, rh, wind, tr, M, W, clo)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Windspeed in meter per second.}

\item{numeric}{tr Air temperature in Celsius degrees.}

\item{numeric}{M Metabolic rate of subject in Watt per mq.}

\item{numeric}{W Work rate of subject in Watt per mq.}

\item{numeric}{clo Clothing insulation level in clo.}
}
\value{

}
\description{
Calculate Predicted Mean Vote (PMV) following ISO 7730.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{PMV_ISO7730}
