% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPD.r
\name{PPD}
\alias{PPD}
\title{PPD}
\usage{
PPD(PMV)
}
\arguments{
\item{PMV}{numeric Predicted Mean Vote.}
}
\value{
Predicted Percentage of Dissatisfied
}
\description{
Predicted Percentage of Dissatisfied (PPD) predicts the percentage of occupants that will be dissatisfied with the thermal conditions.
}
\references{
UNI EN ISO 7730.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{PPD}
