% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvap.r
\name{pvap}
\alias{pvap}
\title{pvap}
\usage{
pvap(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
vapour pressure in hPa.
}
\description{
Vapour pressure estimate from relative humidity.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{pressure}
\keyword{vapour}
