% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVS.r
\name{PVS}
\alias{PVS}
\title{PVS}
\usage{
PVS(T)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degree.}
}
\value{
Saturaion vapor pressure in hPa
}
\description{
Broad range formula to estimate saturation vapor pressure.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Hardy B, 1998,"ITS-90 Formulations for Vapor Pressure, Frostpoint Temperature,Dewpoint Temperature, and Enhancement Factors in the Range 100 to +100 C".Proceedings of the Third International Symposium on Humidity & Moisture",Teddington, London, England, April 1998.
}
\keyword{PVS}

