% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtheoricR.r
\name{radtheoricR}
\alias{radtheoricR}
\title{radtheoricR}
\usage{
radtheoricR(timenow, lat, lon, albedo)
}
\arguments{
\item{timenow}{time "POSIXct" or "POSIXt" object.}

\item{lat}{numeric latitude in decimal degrees.}

\item{lon}{numeric longitude in decimal degrees.}

\item{albedo}{numeric albedo of surface.}
}
\value{
Theoretical radiation in W/mq
}
\description{
R function to assess raw  solar irradiance parameter.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
