% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtodeg.r
\name{radtodeg}
\alias{radtodeg}
\title{radtodeg}
\usage{
radtodeg(anglerad)
}
\arguments{
\item{anglerad}{numeric  Angle in radiands.}
}
\value{
angledeg in degrees.
}
\description{
Conversion angles from radians of an angle in decimal degrees
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{radtodeg}
