% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ral_unacclimatized.r
\name{ral_unacclimatized}
\alias{ral_unacclimatized}
\title{ral_unacclimatized}
\usage{
ral_unacclimatized(met)
}
\arguments{
\item{met}{numeric Metabolic rate in Watt following level of ISO FDIS 7243.}
}
\value{
RAL
}
\description{
Calculate treshshold of risk realated to wbgt index for unacclimatized people.
}
\references{
Ergonomics of the thermal environment – Assessment of heat stress using the WBGT (wet bulb globe temperature)   index, ISO FDIS 7243 (2016)
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{RAL}
