% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_acclimatized.r
\name{rel_acclimatized}
\alias{rel_acclimatized}
\title{rel_acclimatized}
\usage{
rel_acclimatized(met)
}
\arguments{
\item{met}{numeric Metabolic rate in w/mq following level of ISO FDIS 7243.#}
}
\value{
REL
}
\description{
Calculate treshshold of risk realated to wbgt index for acclimatized people
}
\references{
Ergonomics of the thermal environment – Assessment of heat stress using the WBGT (wet bulb globe temperature)   index, ISO FDIS 7243 (2016)
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{REL}
