% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSI_index.r
\name{RSI_index}
\alias{RSI_index}
\title{RSI_index}
\usage{
RSI_index(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
RSI index
}
\description{
Calculates Relative Strain Index.
}
\references{
Lee  DHK  &  Henschel A,1963, Evaluation  of  thermal environment  in  shelters.Rep.  TR 8,  Div.  Occup.  Health,US Dept. Health, Education and Welfare.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
