% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT.r
\name{RT}
\alias{RT}
\title{RT}
\usage{
RT(Ta, rh, v, Tr, M, W, Icl, p, w)
}
\arguments{
\item{numeric}{Ta Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{v Windspeed in meter per second [m/s].}

\item{numeric}{tr Mean radiant temperature in Celsius degree.}

\item{numeric}{M Metabolic work in W/mq of the subject.}

\item{numeric}{W Mechanical work rate in W/mq of the subject.}

\item{numeric}{Icl Available basic clothing insulation [clo]}

\item{numeric}{p  Air permeability of clothing ensemble [lt/m2s]}

\item{numeric}{w  Walking speed in meter per seconds.}
}
\value{
RT
}
\description{
Computes the value of Required Recovery Time following scheme adopted in ISO 11079.Valid under 10 Celsius degree.
}
\author{
Ingvar Holmer. R adaptation Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
ISO 11079, 2007-12-15, ERGONOMICS OF THE THERMAL ENVIRONMENT - DETERMINATION AND INTERPRETATION OF COLD STRESS WHEN USING REQUIRED CLOTHING INSULATION (IREQ) AND LOCAL COOLING EFFECTS.
}
\keyword{Cold}
\keyword{Recovery}
\keyword{Time,}
\keyword{environements.}

