% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharlau_index_delta.r
\name{sharlau_index_delta}
\alias{sharlau_index_delta}
\title{sharlau_index_delta}
\usage{
sharlau_index_delta(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
sharlau_index_delta
}
\description{
Computes the critical sharlau delta.
}
\references{
Scharlau, K., 1950, Einfuhrung eines Schwulemasstabes und Abgrenzung von Schwulezonen durch Isohygrothermen, Erdkunde, 4, 188 201.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{sharlau_index_delta}
