% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharlau_index_state.r
\name{sharlau_index_state}
\alias{sharlau_index_state}
\title{sharlau_index_state}
\usage{
sharlau_index_state(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{
Sharlau State.
}
\description{
State of comfort indicated by Sharlau Index.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Scharlau, K., 1950, Einführung eines Schwülemasstabes und Abgrenzung von Schwülezonen durch Isohygrothermen, Erdkunde, 4, 188-201.
}
\keyword{Sharlau}
\keyword{index}

