% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvol.r
\name{spvol}
\alias{spvol}
\title{spvol}
\usage{
spvol(t, hum_ratio, pa)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees (dry bulb temeprature).}

\item{numeric}{hum_ratio Humidity ratio [unitless].}

\item{numeric}{pa Air pressure in hPa [millibar].}
}
\value{
Specific volume.
}
\description{
Gets the specific volume of air/water vapor mixture [m^3/kg].
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{spvol}
