% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssi_index.r
\name{ssi_index}
\alias{ssi_index}
\title{ssi_index}
\usage{
ssi_index(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in Celsius degrees.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
Summer Simmer index in degC.
}
\description{
Calculate Summer Simmer index developed in Canada.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{SSI}
