% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_class.r
\name{steadman_class}
\alias{steadman_class}
\title{steadman_class}
\usage{
steadman_class(steadman_index)
}
\arguments{
\item{steadman_index}{numeric  Steadman index value.}
}
\value{
class of apparent temperature.
}
\description{
Computes the correspondent Steadman's apparent temperature class.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{apparent}
\keyword{class,}
\keyword{temperature.}
