% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_outdoor_sun.r
\name{steadman_outdoor_sun}
\alias{steadman_outdoor_sun}
\title{steadman_outdoor_sun}
\usage{
steadman_outdoor_sun(t, rh, wind, rshort, sunelev)
}
\arguments{
\item{rshort}{Global short radiation in W/mq.}

\item{sunelev}{Sun elevation in degrees.}

\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Wind speed in meteronseconds ms}
}
\value{

}
\description{
Computes Steadman's apparent temperature under sunny outdoor conditions.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Apparente}
\keyword{Steadman,}
\keyword{temperature}
