% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_shade_class.r
\name{sto_shade_class}
\alias{sto_shade_class}
\title{sto_shade_class}
\usage{
sto_shade_class(t, rh, wind)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Wind speed in meter per seconds [m/s]}
}
\value{
Class of Apparent temperature.
}
\description{
Computes Steadman's apparent temperature class in shade conditions.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{class,shade,apparent}
\keyword{temperature}
