% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_data.r
\name{sun_data}
\alias{sun_data}
\title{sun_data}
\usage{
sun_data(datetime, lat, lon, parameter = "azimuth", tz = "GMT")
}
\arguments{
\item{datetime}{Datetime as in YYYY-MM-DDTHH:MM:SS format IS08601 format.}

\item{numeric}{lat Latitude in decimal degrees.}

\item{numeric}{lon Longitude in decimal degrees.}

\item{character}{parameter Six solar parameter are available by name "azimuth","zenith","solarZenith","elevation","declination","JD".}
}
\value{
value of parameter indicated.
}
\description{
Calculate solar parameter for a location and a time.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{parameters}
\keyword{sun}
