% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_comfort.r
\name{ta_comfort}
\alias{ta_comfort}
\title{ta_comfort}
\usage{
ta_comfort(rh, wind, M, iclo, H)
}
\arguments{
\item{M}{Metabolic rate of the subject in W/mq.}

\item{iclo}{Clothing level of subject in clo unit.}

\item{H}{Timimg in hours.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Wind speed in meter per seconds [m/s].}
}
\value{

}
\description{
Calculate the temperature of thermal comfort requireed given other parameters and timing in hours.
}
\references{
Huizenga
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{ta_comfort}
