% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temprad.r
\name{temprad}
\alias{temprad}
\title{temprad}
\usage{
temprad(t, rh, rshort, rdiffuse, sunelev, albedo = 0.3)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{rshort Direct short-wave solar radiation in Watt on mq.}

\item{numeric}{rdiffuse Diffuse short-wave solar radiation in Watt on mq.}

\item{numeric}{sunelev Sun elevation in degrees.}

\item{numeric}{albedo Surrounding albedo. Default is 0.3.}
}
\value{

}
\description{
Calculate approximate mean radiant temperature.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{temprad}
