% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THI.r
\name{THI}
\alias{THI}
\title{THI}
\usage{
THI(t, rh)
}
\arguments{
\item{t}{numeric  Air temperature in Celsius degrees.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
Temperature Humidity Index discomfort
}
\description{
Given air temperature (Celsius), relative humidity  gives Temperature Humidity Index discomfort. Giles (1990).
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{Humidity}
\keyword{Index}
\keyword{THI,}
\keyword{Temperature}
