% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THI_kliber.r
\name{THI_kibler}
\alias{THI_kibler}
\title{THI_kibler}
\usage{
THI_kibler(t, rh)
}
\arguments{
\item{t}{numeric  Air temperature in Celsius degrees.}

\item{rh}{numeric  Relative humidity in percentage.}
}
\value{
thi index
}
\description{
Calculate Temperature Humidity Index following Kibler equantions. Adopted by Sardinia ARPA.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\references{
Kibler, H.H. Thermal effects of various temperature-humidity combinations on Holstein cattle as measured by eight physiological responses. Res. Bull. 1964, 862, 1–42. 
}
