% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THI_kliber.r
\name{THI_kliber}
\alias{THI_kliber}
\title{THI_kliber}
\usage{
THI_kliber(t, rh)
}
\arguments{
\item{t}{numeric  Air temperature in Celsius degrees.}

\item{rh}{numeric  Relative humidity in percentage.}
}
\value{
thi index
}
\description{
Calculate Temperature Humidity Index following Kliber equantions. Adopted by Sardinia ARPA.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
