% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thom.r
\name{thom}
\alias{thom}
\title{thom}
\usage{
thom(t, rh, wind = 0.2, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Wind speed iq.}

\item{pair}{numeric Air Pressure in hPa. Default is 1010.}
}
\value{
thom index in degC
}
\description{
Calculate Thom discomfort index by using wet bulb temperature estimation.
}
\references{
Thom, E.C. 1959 The discomfort index. Weatherwise, 12, 57,60
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
