% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torr2p.r
\name{torr2p}
\alias{torr2p}
\title{torr2p}
\usage{
torr2p(vp, vpunits = "mbar")
}
\arguments{
\item{numeric}{vp pressure in torr.}

\item{character}{vpunits Unit available are : "atm","bar","Pa","mbar" (default),"dyne/cm2" and "psi".}
}
\value{

}
\description{
Convert pressure in torr in other unit.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{torr2p}

