% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twetbulb_j.r
\name{twetbulb_j}
\alias{twetbulb_j}
\title{twetbulb_j}
\usage{
twetbulb_j(t, p)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{p Air pressure in millibar [hPa].}
}
\value{
wet bulb temperature
}
\description{
Estimating wet bulb temperature following  empirical equation given by Jeevananda, 1976.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{bulb}
\keyword{temperature}
\keyword{wet}
