% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twetbulb_s.r
\name{twetbulb_s}
\alias{twetbulb_s}
\title{twetbulb_s}
\usage{
twetbulb_s(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees. Valid for air pressure 1 atm=101.325 hPa [mb] 760 torr or 760 mmHg.}

\item{numeric}{rh Relative humidity in %.}
}
\value{

}
\description{
Estimate wet bulb temperature by using STULL's empirical equation.
}
\references{
Wet-Bulb Temperature from Relative Humidity and Air Temperature Ronald Stull \link{http://journals.ametsoc.org/doi/pdf/10.1175/JAMC-D-11-0143.1}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{bulb}
\keyword{stull,}
\keyword{temperature,}
\keyword{wet}
