% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbdt.r
\name{wbdt}
\alias{wbdt}
\title{wbdt}
\usage{
wbdt(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Given t air temperature (Celsius), rh relative humidity (percentage)  gives  Wet-bulb dry temperature (WBDT).
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{WBDT}
