% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt.r
\name{wbgt}
\alias{wbgt}
\title{wbgt}
\usage{
wbgt(t, rh, tg)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{tg Globe temperature in Celsius degrees.}
}
\value{

}
\description{
Calculate wet bulb globe temperature index for outdoor environements.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{wbgt}
