% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_stull.r
\name{wbgt_full}
\alias{wbgt_full}
\title{wbgt_stull}
\usage{
wbgt_full(t, rh, tg, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage (%).}

\item{tg}{numeric Solar Radiation Radiance in Watt/mq.}

\item{pair}{numeric Air Pressure in hPa. Default is 1010.}
}
\value{

}
\description{
Calculate wet bulb globe temperature index for outdoor environements 
by using Stull formulation by using air temperature in degC ,relative humidity (%), 
globometric temperature in degC and air pressure in hPa.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{,}
\keyword{Stull}
\keyword{wbgt}
