% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_indoor.r
\name{wbgt_indoor}
\alias{wbgt_indoor}
\title{wbgt_indoor}
\usage{
wbgt_indoor(t, rh, wind = 0.1, press = NULL, topo = NULL)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Air wind wind velocity in meter per second. Default is 0.1.}

\item{numeric}{press Air Pressure at sea level in hPa. Default is 1013.25 hPa.}

\item{numeric}{topo Elevation in meters on sea level. Default is 0.}
}
\value{

}
\description{
Calculate web bulb globe index for indoor environments.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{wbgt_indoor}
