% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_shade.r
\name{wbgt_shade}
\alias{wbgt_shade}
\title{wbgt_shade}
\usage{
wbgt_shade(t, rh, wind, solar = 10, pair = 1010, elev = 0)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Mean Wind speed in meters per second.}

\item{solar}{numeric  Solar Radiation Radiance in Watt./mq}

\item{pair}{numeric Air pressure in millibar or hPa. Default 1010 hPa.}

\item{elev}{numeric Elevation in meters on sea level. Default is 0.}
}
\value{

}
\description{
Calculate wet bulb globe temperature index for outdoor shaded environements.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{shade}
\keyword{wbgt,}
