% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windchill.r
\name{windchill}
\alias{windchill}
\title{windchill}
\usage{
windchill(t, wind)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{wind}{numeric Windspeed in meters per second.}
}
\value{
windchill index
}
\description{
Calculates the NWS Windchill Temperature (WCT) to estimate the perceived temperature with wind in cold environemnt.
}
\references{
Windchill NOAA calculator \url{https://www.weather.gov/epz/wxcalc_windchill}
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{degC.}
\keyword{in}
\keyword{index}
\keyword{windchill}
