% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windchill_cla.r
\name{windchill_cla}
\alias{windchill_cla}
\title{windchill_cla}
\usage{
windchill_cla(t, wind)
}
\arguments{
\item{t}{numeric    Air temperature in degC.}

\item{wind}{numeric Windspeed in meters per second.}
}
\value{
windchill index in degC
}
\description{
Calculates the risk class of Windchill index.
}
\references{
Windchill Tables \url{https://www.canada.ca/en/environment-climate-change/services/weather-health/wind-chill-cold-weather/wind-chill-index.html}
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{classes}
\keyword{windchill}
