% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winddir.r
\name{winddir}
\alias{winddir}
\title{winddir}
\usage{
winddir(u, v)
}
\arguments{
\item{u}{numeric U zonal component}

\item{v}{numeric V meridian component}
}
\description{
Calculate meteorological wind direction from cartesian components u and v.
}
\references{
Istituto di Biometeorologia Firenze Italy
}
\author{
Alfonso crisci \email{a.crisci@ibimet.cnr.it} Marco Morabito \email{m.morabito@unifi.it}
}
\keyword{Meteorological}
\keyword{Wind}
\keyword{direction}
