% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windspeed.r
\name{windspeed}
\alias{windspeed}
\title{windspeed}
\usage{
windspeed(u, v)
}
\arguments{
\item{u}{numeric U zonal component}

\item{v}{numeric V meridian component}
}
\value{

}
\description{
Calculate meteorological wind speed.
}
\references{
Istituto di Biometeorologia Firenze Italy.
}
\author{
Alfonso crisci \email{a.crisci@ibimet.cnr.it} Marco Morabito \email{m.morabito@unifi.it}
}
\keyword{speed}
\keyword{wind}
