% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancePMV7730.r
\name{balancePMV7730}
\alias{balancePMV7730}
\title{balancePMV7730}
\usage{
balancePMV7730(pmv, M)
}
\arguments{
\item{pmv}{numeric Predicted Mean Vote Fanger following ISO 7730.}

\item{M}{numeric Metabolism of the person (met).}
}
\value{

}
\description{
Calculate the thermal balance of the body in W/mq.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{balancePMV7730}
