% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2F.r
\name{C2F}
\alias{C2F}
\title{C2F}
\usage{
C2F(C)
}
\arguments{
\item{numeric}{C Temperature in Celsius degrees.}
}
\value{

}
\description{
Convert temperature from Celsius degree to Fahrenheit degree.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{temperature}
