% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clomin_custom.r
\name{clomin_custom}
\alias{clomin_custom}
\title{clomin_custom}
\usage{
clomin_custom(t, rh, wind, trad, age, mbody, ht, gender = "male")
}
\arguments{
\item{t}{numeric        Air temperature in degC.}

\item{rh}{numeric       Air Relative humidity in percentage.}

\item{wind}{numeric     Wind speed in meters per second.}

\item{age}{numeric    Age in years.}

\item{mbody}{numeric    Body Mass in kg.}

\item{ht}{numeric       Heigth of subject in meters.}

\item{gender}{character Sex of the subject. Default is "male" otherwise "female".}

\item{tr}{numeric       Mean radiant temperature in degC.}

\item{M}{numeric        Metabolic rate of subject in W/mq.}

\item{W}{numeric        Mechanical work of subject in W/mq. Generally 0.}
}
\value{

}
\description{
Calculate minimal custom clothing insulation value required to reach thermal comfort in moderate thermal environments ( -0.5 < PMV <0.5) based on PMV ISO 7730.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{ocir}
