% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_16.r
\name{compass_16}
\alias{compass_16}
\title{compass_16}
\usage{
compass_16(direction)
}
\arguments{
\item{direction}{numeric   Directions in degrees from the North.}
}
\value{

}
\description{
Calculate the wind sector (16 partition) provenience.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{compass_16}
