% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_8.r
\name{compass_8}
\alias{compass_8}
\title{compass_8}
\usage{
compass_8(direction)
}
\arguments{
\item{numeric}{direction Directions in degrees from the North.}
}
\value{

}
\description{
Calculate the  wind sector (8 partition)  of its provenience.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{compass_8}
