% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayname_IT.r
\name{dayname_IT}
\alias{dayname_IT}
\title{dayname_IT}
\usage{
dayname_IT(date)
}
\arguments{
\item{date}{Date in YYYY-MM-DD format.}
}
\value{

}
\description{
Give the name of day in italian languages.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{day.}
\keyword{italian}
\keyword{name}
\keyword{of}
