% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deficitsat.r
\name{deficitsat}
\alias{deficitsat}
\title{deficitsat}
\usage{
deficitsat(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Computes saturation deficit in hPa. This is air dryness evaluation.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{deficit}
\keyword{saturation}

