% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degToRad.r
\name{degToRad}
\alias{degToRad}
\title{degToRad}
\usage{
degToRad(angleDeg)
}
\arguments{
\item{numeric}{angleDeg Angle in decimal degrees.}
}
\value{

}
\description{
Conversion of decimal degrees to radians.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{degToRad}

