% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewpoint.r
\name{dewpoint}
\alias{dewpoint}
\title{dewpoint}
\usage{
dewpoint(t, rh, formula = "NOAA")
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{character}{formula Type of formula used for saturation pressure : "NOAA","Sonntag","Paroscientific".Default is "NOAA".}
}
\value{

}
\description{
Computes Dewpoint temperature in Celsius degree following differentschemes.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{dewpoint}

