% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusivity.r
\name{diffusivity}
\alias{diffusivity}
\title{diffusivity}
\usage{
diffusivity(t, Pair)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{Pair Air pressure in Millibar [hPa]}
}
\value{

}
\description{
Calculate air diffusivity.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{diffusivity}

