% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discomfort_index.r
\name{discomfort_index}
\alias{discomfort_index}
\title{discomfort_index}
\usage{
discomfort_index(t, rh, wind = 0.2, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meters per second.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{
discomfort_index
}
\description{
Given t air temperature (degC), rh relative humidity , wind speed in meters per second and air pressure ( hPa) gives  Discomfort index (DI).
}
\references{
Thom, E.C., 1959, The discomfort index. Weatherwise, 12, 57 60.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
