% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_atm.r
\name{emis_atm}
\alias{emis_atm}
\title{emis_atm}
\usage{
emis_atm(tk, rh)
}
\arguments{
\item{tk}{numeric Air temperature in degK.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
air emissivity
}
\description{
Calculate air emissivity.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{emis_atm}
