% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_atm.r
\name{emis_atm}
\alias{emis_atm}
\title{emis_atm}
\usage{
emis_atm(ta, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Calculate air emissivity.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{emis_atm}
