% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enthalpy.r
\name{enthalpy}
\alias{enthalpy}
\title{enthalpy}
\usage{
enthalpy(t, hum_ratio)
}
\arguments{
\item{hum_ratio}{Humidity ratio [unitless].}

\item{numeric}{t Air temperature in Celsius degrees.}
}
\value{

}
\description{
Calculates the enthalpy of air/water vapor mixture [kJ/kg]. Inputs are dry-bulb temperature [C] and humidity ratio [unitless].
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{enthalpy}
