% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2C.r
\name{F2C}
\alias{F2C}
\title{F2C}
\usage{
F2C(F)
}
\arguments{
\item{numeric}{F Temperature in Fahrenheit degrees}
}
\value{

}
\description{
Convert temperature from Fahrenheit degree to  Celsius degree.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{F2C}
