% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits_index.r
\name{fits_index}
\alias{fits_index}
\title{fits_index}
\usage{
fits_index(t, rh, wind = 0.2, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meters per second.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{
fits index
}
\description{
Calculate Fighter Index of Thermal Stress FITS.
}
\references{
Nunneley SH, Stribley F, 1979, Fighter index of thermal stress FITS  guidance for hot-weather aircraft operations.Aviat Space Environ Med 50, 639 42.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
