% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRH.r
\name{fRH}
\alias{fRH}
\title{fRH}
\usage{
fRH(t, td)
}
\arguments{
\item{t}{numeric Vector or value of air temperature in Celsius degree.}

\item{td}{numeric Vector or value of dew point temperature in Celsius degree.}
}
\value{
relative humidity value
}
\description{
Return relative humidity from air temperature (Celsius) and Dew point (Celsius);
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{Relative}
\keyword{humidity}
