% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRH.r
\name{fRH}
\alias{fRH}
\title{fRH}
\usage{
fRH(tair, dewp)
}
\arguments{
\item{tair}{numeric Vector or value of air temperature in Celsius degree.}

\item{dewp}{numeric Vector or value of dew point temperature in Celsius degree.}
}
\value{
res numeric  relative humidity value
}
\description{
Return relative humidity from air temperature (Celsius) and Dew point (Celsius);
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Relative}
\keyword{humidity}

