% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frostime.r
\name{frostime}
\alias{frostime}
\title{frostime}
\usage{
frostime(t, wind)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{wind The wind speed in meter per seconds [m/s].}
}
\value{

}
\description{
Given a temperature t (Celsius) and wind ( m/sec) frost time following Wind Chill Frostbite Chart.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
ISO 9920:2007 Ergonomics of the thermal environment -- Estimation of thermal insulation and water vapour resistance of a clothing ensemble.
}
\keyword{frostime}

