% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts2kmh.r
\name{fts2kmh}
\alias{fts2kmh}
\title{fts2kmh}
\usage{
fts2kmh(fts)
}
\arguments{
\item{numeric}{fts Speed in feet per second.}
}
\value{

}
\description{
Conversion speed in feet per second to kilometer per hour.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{fts2kmh}

