% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts2knots.r
\name{fts2knots}
\alias{fts2knots}
\title{fts2knots}
\usage{
fts2knots(fts)
}
\arguments{
\item{numeric}{fts Speed in feet per second.}
}
\value{

}
\description{
Conversion from feet per second to knot per second.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{fts2knots}

