% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDOY.r
\name{getDOY}
\alias{getDOY}
\title{getDOY}
\usage{
getDOY(datetime)
}
\arguments{
\item{date}{Date in format YYYY-MM-DD.}
}
\value{
numeric Julian day.
}
\description{
Get julian day number of date.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{getDOY}

