% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_index.r
\name{heat_index}
\alias{heat_index}
\title{heat_index}
\usage{
heat_index(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage (%).}
}
\value{
heat index in degC.
}
\description{
Calculate the heat index following the NOAA scheme.
}
\references{
NOAA index from George Winterling based on SteadMan's work "The Assessment of Sultriness. Part I: A Temperature-Humidity Index Based on Human Physiology and Clothing Science, R. G. Steadman, Journal of Applied Meteorology, July 1979, Vol 18 No7, pp 861 873.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{NOAA}
\keyword{heat}
\keyword{index}
