% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_index.r
\name{hi_index}
\alias{hi_index}
\title{hi_index}
\usage{
hi_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Given air temperature (Celsius), relative humidity (%) give a heat stress index in Celsius degree for primates animals.
}
\references{
Hill et al (2004), Indices of environmental temperatures for primates in open habitats. Websource \link(https://community.dur.ac.uk/r.a.hill/Hill%20et%20al%202004.pdf)
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{hi_index}
