% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSI_index.r
\name{HSI_index}
\alias{HSI_index}
\title{HSI_index}
\usage{
HSI_index(t, rh, wind, tr, M, W, clo, param = "HSI")
}
\arguments{
\item{t}{numeric    Air temperature in degC.}

\item{rh}{numeric   Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meter per second.}

\item{tr}{numeric Mean radiant temperature in degC.}

\item{M}{numeric    Metabolic work in Watt per mq of the subject.}

\item{W}{numeric  Mechanical work rate in Watt per mq of the subject.}

\item{clo}{numeric  Available basic clothing insulation in clo.}

\item{param}{character Index requested HSI or AET Allowable exposure time in minutes.}
}
\value{
HSI index
}
\description{
Computes the Heat Strain Index. HSI is the ratio of the demand for sweat evaporation to capacity of evaporation (Ereq on Emax). 
This denotes also the percent of skin wettedness, which is a good predictor of warm discomfort.
}
\references{
Belding et al 1955. Index for evaluating heat stress in terms of resulting physiological strain. Heat Pip Air Condit 27, 129 36.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{HSI}
