% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidex.r
\name{humidex}
\alias{humidex}
\title{humidex}
\usage{
humidex(t, rh)
}
\arguments{
\item{t}{numeric  Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}
}
\value{
humidex in degC.
}
\description{
Computes canadian Humidex index.
}
\references{
Masterson, J. et Richardson, F. A., Humidex, A Method of Quantifying Human Discomfort Due to Excessive Heat and Humidity, Downsview, Ontario, Environnement Canada, 1979, 45.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{canada}
\keyword{humidex,}
