% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humrat.r
\name{humrat}
\alias{humrat}
\title{humrat}
\usage{
humrat(p, pa)
}
\arguments{
\item{numeric}{p Water pressure in hPa.}

\item{numeric}{pa Air pressure in hPa.}
}
\value{

}
\description{
Calculates humidity ratio in vapor-air mixture.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{humrat}
