% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humrat.r
\name{humrat}
\alias{humrat}
\title{humrat}
\usage{
humrat(p, pair)
}
\arguments{
\item{p}{numeric  Water pressure in hPa.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{

}
\description{
Calculates humidity ratio in vapor-air mixture.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{humrat}
