% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IREQ.r
\name{IREQ}
\alias{IREQ}
\title{IREQ}
\usage{
IREQ(t, rh, wind, tr, M, W, clo, p, w, param = "IREQ_neutral")
}
\arguments{
\item{t}{numeric    Air temperature in degC.}

\item{rh}{numeric   Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meter per second.}

\item{tr}{numeric Mean radiant temperature in degC.}

\item{M}{numeric    Metabolic work in Watt per mq of the subject.}

\item{W}{numeric  Mechanical work rate in Watt per mq of the subject.}

\item{clo}{numeric  Available basic clothing insulation in clo.}

\item{p}{numeric   Air permeability of clothing ensemble lt on meters per second.}

\item{w}{numeric   Walking speed in meters per second.}

\item{param}{character Parameter of IREQ ISO 11079: "IREQ_neutral","IREQ_min","ICL_neutral","ICL_min","DLE_neutral","DLE_min","ICL_ISO11079"}
}
\value{
IREQ value
}
\description{
Calculation of Required Clothing Insulation IREQ.
}
\references{
ISO 11079, 2007-12-15, ERGONOMICS OF THE THERMAL ENVIRONMENT - DETERMINATION AND INTERPRETATION OF COLD STRESS WHEN USING REQUIRED CLOTHING INSULATION (IREQ) AND LOCAL COOLING EFFECTS.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{IREQ}
