% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IREQ.r
\name{IREQ}
\alias{IREQ}
\title{IREQ}
\usage{
IREQ(t, rh, v, tr, M, W, Icl, p, w, param = "IREQ_neutral")
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{v  Relative air velocity in meter per seconds.}

\item{numeric}{tr Mean radiant temperature in Celsius degree.}

\item{numeric}{M  Metabolic energy production [W/mq].}

\item{numeric}{W Rate of mechanical work [W/mq].}

\item{numeric}{Icl Available basic clothing insulation [clo]}

\item{numeric}{p  Air permeability of clothing ensemble [lt/m2s]}

\item{numeric}{w  Walking speed in meter per seconds.}

\item{character}{param  Parameter of IREQ ISO 11079: "IREQ_neutral","IREQ_min","ICL_neutral","ICL_min","DLE_neutral","DLE_min","ICL_ISO11079"}
}
\value{
IREQ value
}
\description{
Calculation of Required Clothing Insulation (IREQ) following elaboration of http://www.eat.lth.se/fileadmin/eat/Termisk_miljoe/IREQ2009ver4_2.html
}
\author{
Ingvar Holmer \link{http://www.eat.lth.se/fileadmin/eat/Termisk_miljoe/IREQ2009ver4_2.htm}. R adaptation Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
ISO 11079, 2007-12-15, ERGONOMICS OF THE THERMAL ENVIRONMENT - DETERMINATION AND INTERPRETATION OF COLD STRESS WHEN USING REQUIRED CLOTHING INSULATION (IREQ) AND LOCAL COOLING EFFECTS.
}
\keyword{IREQ}

