% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLeapYear.r
\name{isLeapYear}
\alias{isLeapYear}
\title{isLeapYear}
\usage{
isLeapYear(yr)
}
\arguments{
\item{numeric}{yr Year}
}
\value{

}
\description{
Logical if the year indicated is a leap year.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{isLeapYear}

