% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2fts.r
\name{kmh2fts}
\alias{kmh2fts}
\title{kmh2fts}
\usage{
kmh2fts(kmh)
}
\arguments{
\item{numeric}{kmh Speed in kilometer per hour.}
}
\value{

}
\description{
Conversion from kilometer per hour to feet per second.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{kmh2fts}

