% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2mph.r
\name{kmh2mph}
\alias{kmh2mph}
\title{kmh2mph}
\usage{
kmh2mph(kmh)
}
\arguments{
\item{numeric}{kmh Speed in Kilometer per hours.}
}
\value{

}
\description{
Conversion kilometer per hour to mile per hour.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{kmh2mph}
