% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmh2ms.r
\name{kmh2ms}
\alias{kmh2ms}
\title{kmh2ms}
\usage{
kmh2ms(kmh)
}
\arguments{
\item{kmh}{numeric  Speed kilometer per hour.}
}
\value{

}
\description{
Conversion from kilometer per hour to meter per second.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{kmh2ms}
