% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2fts.r
\name{knots2fts}
\alias{knots2fts}
\title{knots2fts}
\usage{
knots2fts(knots)
}
\arguments{
\item{knots}{numeric  Speed in knots per second.}
}
\value{
feets per second
}
\description{
Conversion from knots per second to feets per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{knots2fts}
