% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2fts.r
\name{knots2fts}
\alias{knots2fts}
\title{knots2fts}
\usage{
knots2fts(knots)
}
\arguments{
\item{numeric}{knots Speed in knot per second.}
}
\value{

}
\description{
Conversion from knot per second to feet per second.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{knots2fts}

