% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2kmh.r
\name{knots2kmh}
\alias{knots2kmh}
\title{knots2kmh}
\usage{
knots2kmh(knots)
}
\arguments{
\item{knots}{numeric  Speed in knots per second.}
}
\value{
kilometers per hour
}
\description{
Conversion from knot per second to kilometers per hour.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{knots2kmh}
