% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2mph.r
\name{knots2mph}
\alias{knots2mph}
\title{knots2mph}
\usage{
knots2mph(knots)
}
\arguments{
\item{knots}{numeric  Speed in knots per second.}
}
\value{

}
\description{
Conversion from knot per second to mile per hour.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{knots2mph}
