% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2ms.r
\name{knots2ms}
\alias{knots2ms}
\title{knots2ms}
\usage{
knots2ms(knots)
}
\arguments{
\item{knots}{numeric  Speed in knots per second.}
}
\value{

}
\description{
Conversion from feet per second to meter per second.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{knots2ms}
