% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtolbsinch.r
\name{mbtolbsinch}
\alias{mbtolbsinch}
\title{mbtolbsinch}
\usage{
mbtolbsinch(mb)
}
\arguments{
\item{numeric}{mb Air pressure in Millibar [hPa].}
}
\value{

}
\description{
Conversion from Millibar [hPa] to inches.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{mbtolbsinch}
