% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtommHG.r
\name{mbtommHG}
\alias{mbtommHG}
\title{mbtommHG}
\usage{
mbtommHG(mb)
}
\arguments{
\item{mb}{numeric   Air pressure in Millibar [hPa].}
}
\value{
mmHG
}
\description{
Conversion from Millibar [hPa] to mmHG.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{mbtommHG}
