% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdi_index.r
\name{mdi_index}
\alias{mdi_index}
\title{mdi_index}
\usage{
mdi_index(t, rh, wind = 0.2, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meters per second.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{
mdi_index
}
\description{
Modified discomfort index MDI.
}
\references{
Moran DS, Shitzer A, Pandolf KB , 1998, A physiological strain index to evaluate heat stress. Am J Physiol 275, R 129 34.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{mdi_index}
