% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2fts.r
\name{mph2fts}
\alias{mph2fts}
\title{mph2fts}
\usage{
mph2fts(mph)
}
\arguments{
\item{mph}{numeric  Speed miles per hour.}
}
\value{

}
\description{
Conversion speed in  miles per hour to feet per second.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{mph2fts}
