% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2kmh.r
\name{mph2kmh}
\alias{mph2kmh}
\title{mph2kmh}
\usage{
mph2kmh(mph)
}
\arguments{
\item{mph}{numeric  Speed miles per hour.}
}
\value{
kilometers per hour
}
\description{
Conversion speed in  miles per hour to kilometers per hour.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{mph2kmh}
