% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2knots.r
\name{mph2knots}
\alias{mph2knots}
\title{mph2knots}
\usage{
mph2knots(mph)
}
\arguments{
\item{numeric}{mph Speed in miles per hour.}
}
\value{

}
\description{
Conversion speed in  miles per hour to knots per second.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{mph2knots}

