% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2ms.r
\name{mph2ms}
\alias{mph2ms}
\title{mph2ms}
\usage{
mph2ms(mph)
}
\arguments{
\item{numeric}{mph Speed in miles on hours.}
}
\value{

}
\description{
Conversion miles per hour in meter per second.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{mph2ms}
