% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mph2ms.r
\name{mph2ms}
\alias{mph2ms}
\title{mph2ms}
\usage{
mph2ms(mph)
}
\arguments{
\item{mph}{numeric  Speed miles per hour.}
}
\value{
meters per second
}
\description{
Conversion miles per hour in meters per second.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{mph2ms}
