% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_globe}
\alias{mrt_globe}
\title{mrt_globe}
\usage{
mrt_globe(t, tg, wind, diam)
}
\arguments{
\item{t}{numeric air temperature in degC.}

\item{tg}{numeric global short solar irradiance in Watt on mq.}

\item{wind}{numeric windspeed in meter per second.}

\item{diam}{numeric diameter of the sphere in millimeters. Input example 50 mm.}
}
\value{
Mean Radiant temperature in degC
}
\description{
Given air temperature , the globe Temeperature, the wind speed and diameter mean radiant temperature is done according ISO 7726 1998.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{globe}
\keyword{temperature}
