% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_solar_proj}
\alias{mrt_solar_proj}
\title{mrt_solar_proj}
\usage{
mrt_solar_proj(t, rh, solar, sunelev, alb_sfc, fdir, emis_sfc)
}
\arguments{
\item{t}{numeric        Air temperature in degC.}

\item{rh}{numeric       Relative humidity in percentage.}

\item{solar}{numeric   Global solar radiation in Watt on mq.}

\item{sunelev}{numeric  Sun elevation angle in decimal degrees.}

\item{alb_sfc}{numeric  Mean albedo of surroundings. Default is 0.4.}

\item{fdir}{numeric    Ratio between directed ad diffuse radiation.}

\item{emis_sfc}{numeric Surface emissivity.}
}
\value{
Mean Radiant temperature projected in degC.
}
\description{
Calculated the mean radiant temperature from the short solar irradiance with human projection factor. it assumes a uniform surround temperature of air temperature  and short wave solar radiation.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{fMRT}
