% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt_functions.r
\name{mrt_thorsson}
\alias{mrt_thorsson}
\title{mrt_thorsson}
\usage{
mrt_thorsson(t, tg, wind, diam)
}
\arguments{
\item{t}{numeric air temperature in degC.}

\item{tg}{numeric global short solar irradiance in Watt on mq.}

\item{wind}{numeric windspeed in meter per second.}

\item{diam}{numeric diameter of the sphere in millimeter. Input example 50 mm.}
}
\value{
Mean Radiant temperature in degC
}
\description{
Calculated the mean radiant temperature from the solar radiation.  Modified based on direct and diffuse ratio. Assumes a uniform surround temperature of Ta && short wave solar radiation only;
}
\references{
Thorsson et al, Different methods for estimating the mean radiant temperature in an outdoor urban setting, Int. J. Climatol.27  1983   1993 (2007)
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{MRT}
