% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2kmh.r
\name{ms2kmh}
\alias{ms2kmh}
\title{ms2kmh}
\usage{
ms2kmh(ms)
}
\arguments{
\item{numeric}{ms Speed in meter per second [m/s].}
}
\value{

}
\description{
Conversion speed meter per second  in km per hours.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{ms2kmh}
