% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2knots.r
\name{ms2knots}
\alias{ms2knots}
\title{ms2knots}
\usage{
ms2knots(ms)
}
\arguments{
\item{numeric}{ms Speed in meter per second [m/s].}
}
\value{

}
\description{
Conversion speed meter per second in knot per second.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{ms2knots}
