% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2mph.r
\name{ms2mph}
\alias{ms2mph}
\title{ms2mph}
\usage{
ms2mph(ms)
}
\arguments{
\item{ms}{numeric Speed in meters per second.}
}
\value{
miles per hour
}
\description{
Conversion speed in meters per second to miles per hour.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{ms2mph}
