% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_index.r
\name{net_index}
\alias{net_index}
\title{net_index}
\usage{
net_index(t, rh, wind)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meters per second.}
}
\value{

}
\description{
Calculate the Net Effective Temperature (NET) incorporating ambient temperature in degC , wind speed (m/s), and relative humidity (%). It is used in  hot and cold situations.
}
\references{
Cho KLA, Margaret Loughnan M, and Tapper T, 2013, An Exploration of Temperature Metrics for Further Developing the Heat-Health Weather Warning System in Hong Kong.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}#' @keywords  net_index
}
