% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_index.r
\name{oxford_index}
\alias{oxford_index}
\title{oxford_index}
\usage{
oxford_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Computes  the oxford index index of discomfort.
}
\references{
Lind AR, Hallon RF,1957, Assessment of physiologic severity of hot climate. J Appl Physiol 11, 35–40.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{oxford_index}
