% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_index.r
\name{oxford_index}
\alias{oxford_index}
\title{oxford_index}
\usage{
oxford_index(t, rh, wind = 0.2, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meters per second.}

\item{pair}{numeric Air pressure in hPa.}
}
\value{
oxford index
}
\description{
Computes the oxford index of thermal discomfort.
}
\references{
Lind AR, Hallon RF,1957, Assessment of physiologic severity of hot climate. J Appl Physiol 11, 35 40.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{oxford_index}
