% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_local.r
\name{p_local}
\alias{p_local}
\title{p_local}
\usage{
p_local(press = 1010, elev, t)
}
\arguments{
\item{press}{numeric Air sea level pressure in hPa.}

\item{elev}{numeric Altitude or Elevation in meters.}

\item{t}{numeric Mean Air temperature in degC.}
}
\value{
local air pressure
}
\description{
Estimate local value of air pressure in hPa.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{p_local}
