% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMV_custom.r
\name{PMV_custom}
\alias{PMV_custom}
\title{PMV_custom}
\usage{
PMV_custom(t, rh, wind, tr, iclo, age, mbody, ht, gender = "male")
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Windspeed in meter per second.}

\item{numeric}{tr Air temperature in Celsius degrees.}

\item{numeric}{M Metabolic rate of subject Watt per mq.}

\item{numeric}{W Work rate of subject in Watt per mq.}

\item{numeric}{clo Clothing insulation level in clo.}

\item{numeric}{age Age in years.}

\item{numeric}{mbody Body Mass in kg.}

\item{numeric}{ht Heigth of subject in meters.}

\item{character}{gender Sex of the subject. Default is "male".}
}
\value{

}
\description{
Calculate Predicted Mean Vote (PMV) following ISO 7730 customizing some individual features.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{PMV_custom}
