% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMV_ISO7730.r
\name{PMV_ISO7730}
\alias{PMV_ISO7730}
\title{PMV_ISO7730}
\usage{
PMV_ISO7730(t, rh, wind, tr, M, W, clo)
}
\arguments{
\item{t}{numeric    Air temperature in degC.}

\item{rh}{numeric   Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meter per second.}

\item{tr}{numeric Mean radiant temperature in degC.}

\item{M}{numeric    Metabolic work in Watt per mq of the subject.}

\item{W}{numeric  Mechanical work rate in Watt per mq of the subject.}

\item{clo}{numeric  Available basic clothing insulation in clo.}
}
\value{
Predicted Mean Vote
}
\description{
Calculate Predicted Mean Vote (PMV) following ISO 7730.
}
\author{
Istituto per la Bioeconomia Firenze CNR Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{PMV_ISO7730}
