% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poda.r
\name{poda}
\alias{poda}
\title{poda}
\usage{
poda(t, rh, p)
}
\arguments{
\item{p}{Local air pressure in hPa [millibar].}

\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Calculate partial pressure of oxygen in function of altitude.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{altitude.}
\keyword{oxygen,}
\keyword{poda,}
