% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poda.r
\name{poda}
\alias{poda}
\title{poda}
\usage{
poda(t, rh, pair)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Air Relative humidity in percentage.}

\item{pair}{numeric Local air pressure in hPa.}
}
\value{
partial pressure of oxygen
}
\description{
Calculate the partial pressure of oxygen in function of altitude.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{altitude.}
\keyword{oxygen,}
\keyword{poda,}
