% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPD.r
\name{PPD}
\alias{PPD}
\title{PPD}
\usage{
PPD(PMV)
}
\arguments{
\item{numeric}{PMV Preadicted Mean Vote}
}
\value{

}
\description{
Predicted Percentage of Dissatisfied (PPD) predicts the percentage of occupants that will be dissatisfied with the thermal conditions.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
UNI EN ISO 7730.
}
\keyword{PPD}

