% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projman.r
\name{projman}
\alias{projman}
\title{projman}
\usage{
projman(sunelev)
}
\arguments{
\item{numeric}{sunelev SUn elevation in degrees.}
}
\value{
Projection factor
}
\description{
Calculate the Human Body Projection Area factor in function to sun elevation.
}
\references{
Jendritzky G (1990) Bioklimatische Bewertungsgrundlage der Raume am Beispiel von. 27 mesoskaligen Bioklimakarten.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Area}
\keyword{factor.}
\keyword{projection}
