% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvap.r
\name{pvap}
\alias{pvap}
\title{pvap}
\usage{
pvap(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{
vapour pressure in hPa.
}
\description{
Vapour pressure estimate from relative humidity.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{pressure}
\keyword{vapour}

