% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVS.r
\name{PVS}
\alias{PVS}
\title{PVS}
\usage{
PVS(T)
}
\arguments{
\item{t}{numeric Air temperature in degC}
}
\value{
Saturaion vapor pressure (hPa)
}
\description{
Broad range formula to estimate saturation vapor pressure.
}
\references{
Hardy B, 1998,"ITS-90 Formulations for Vapor Pressure, Frostpoint Temperature,Dewpoint Temperature, and Enhancement Factors in the Range 100 to +100 C".Proceedings of the Third International Symposium on Humidity & Moisture",Teddington, London, England, April 1998.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{PVS}
