% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_direct_tilted.r
\name{rad_direct_tilted}
\alias{rad_direct_tilted}
\title{rad_direct_tilted}
\usage{
rad_direct_tilted(datetime, lat, lon, planezen, planeaz)
}
\arguments{
\item{planezen}{Zenital angles of the plane.}

\item{planeaz}{Azimuthal angles of the plane.}

\item{date}{Date in format YYYY-MM-DD HH:MM:SS}

\item{numeric}{lat Latitude in decimal degrees.}

\item{numeric}{lon Longitude in decimal degrees.}
}
\value{

}
\description{
Computed a incident tilted radiation for a inclinated (planezen) and orientated plane (planeaz).
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Tilted}
\keyword{radiation.}
