% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtheoric.r
\name{radtheoric}
\alias{radtheoric}
\title{radtheoric}
\usage{
radtheoric(jddate, sunelev = 75, albedo = 0.4, param = "direct")
}
\arguments{
\item{jddate}{Datetime in DOY Day of year}

\item{sunelev}{numeric Sun elevation decimal degrees}

\item{albedo}{numeric Surronding albedo value. Default is 0.3}

\item{param}{character Theoretical parameter available for radiation "global","direct","diffuse" ( clear sky diffuse),"reflected".}
}
\value{
Requested irradinace in KJ on mq
}
\description{
Computes different theoretical radiation in W/mq for a location and a time : Global, Direct Beam radiation,Diffuse Radiation, Reflected radiation based on albedo.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{radtheoric}
