% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtheoric.r
\name{radtheoric}
\alias{radtheoric}
\title{radtheoric}
\usage{
radtheoric(datetime, lat, lon, albedo = 0.3, param = "global")
}
\arguments{
\item{date}{Datetime in format YYYY-MM-DD HH:MM:SS}

\item{numeric}{lat Latitude in decimal degrees.}

\item{numeric}{lon Longitude in decimal degrees.}

\item{numeric}{albedo Surronding albedo value. Default is 0.3.}

\item{character}{param Theoretical parameter available for radiation "global","direct","diffuse","reflected".}
}
\value{

}
\description{
Computes different theoretical radiation in W/mq for a location and a time : Global, Direct Beam radiation,Diffuse Radiation, Reflected radiation based on albedo.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{radtheoric}

