% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radtheoricR.r
\name{radtheoricR}
\alias{radtheoricR}
\title{radtheoricR}
\usage{
radtheoricR(timenow, lat, long, albedo)
}
\arguments{
\item{timenow}{"POSIXct" or "POSIXt" object.}

\item{lat}{latitude in decimal degrees.}

\item{lat}{longitude in decimal degrees.}
}
\value{
theoretical radiation in a location at time.
}
\description{
R function to assess raw  solar irradiance parameter.
}
\references{
Istituto di Biometeorologia Firenze Italy
Centro Interdipartimentale di Bioloclimatologia University of Florence
}
\author{
Alfonso crisci \email{a.crisci@ibimet.cnr.it} Marco Morabito \email{m.morabito@unifi.it}
}
