% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radToDeg.r
\name{radToDeg}
\alias{radToDeg}
\title{radToDeg}
\usage{
radToDeg(angleRad)
}
\arguments{
\item{numeric}{angleRad  Angle in radiands.}
}
\value{
angleDeg in degrees.
}
\description{
Conversion the radians of an angle in decimal degrees.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{radToDeg}

