% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSI_index.r
\name{RSI_index}
\alias{RSI_index}
\title{RSI_index}
\usage{
RSI_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Calculates Relative strain index.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Lee  DHK  &  Henschel A,(1963)  Evaluation  of  thermal environment  in  shelters.Rep.  TR–8,  Div.  Occup.  Health,US Dept. Health, Education and Welfare
}
\keyword{RSI_index}

