% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT.r
\name{RT}
\alias{RT}
\title{RT}
\usage{
RT(t, rh, wind, trad, M, W, clo, p, w)
}
\arguments{
\item{t}{numeric    Air temperature in degC.}

\item{rh}{numeric   Relative humidity in percentage.}

\item{wind}{numeric Windspeed in meter per second.}

\item{trad}{numeric Mean radiant temperature in degC.}

\item{M}{numeric    Metabolic work in Watt per mq of the subject.}

\item{W}{numeric  Mechanical work rate in Watt per mq of the subject.}

\item{clo}{numeric  Available basic clothing insulation in clo.}

\item{p}{numeric   Air permeability of clothing ensemble lt on meters per second.}

\item{w}{numeric   Walking speed in meters per second.}
}
\value{
Required Recovery Time
}
\description{
Computes the value of Required Recovery Time following scheme adopted in ISO 11079. Valid under 10 Celsius degree.
}
\references{
ISO 11079, 2007-12-15, ERGONOMICS OF THE THERMAL ENVIRONMENT - DETERMINATION AND INTERPRETATION OF COLD STRESS WHEN USING REQUIRED CLOTHING INSULATION (IREQ) AND LOCAL COOLING EFFECTS.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}

Ingvar Holmer. R adaptation Istituto di Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{Recovery}
\keyword{Time.}
