% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharlau_index_delta.r
\name{sharlau_index_delta}
\alias{sharlau_index_delta}
\title{sharlau_index_delta}
\usage{
sharlau_index_delta(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{

}
\description{
Computes the critical sharlau delta in Celsius degrees.
}
\references{
Scharlau, K., 1950, Einfuhrung eines Schwulemasstabes und Abgrenzung von Schwulezonen durch Isohygrothermen, Erdkunde, 4, 188 201.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{sharlau_index_delta}
