% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharlau_index_delta.r
\name{sharlau_index_delta}
\alias{sharlau_index_delta}
\title{sharlau_index_delta}
\usage{
sharlau_index_delta(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Computes the critical sharlau delta in Celsius degrees.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Scharlau, K., 1950, Einführung eines Schwülemasstabes und Abgrenzung von Schwülezonen durch Isohygrothermen, Erdkunde, 4, 188-201.
}
\keyword{sharlau_index_delta}

