% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharlau_index_state.r
\name{sharlau_index_state}
\alias{sharlau_index_state}
\title{sharlau_index_state}
\usage{
sharlau_index_state(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{
Sharlau State.
}
\description{
State of comfort indicated by Sharlau Index delta.
}
\references{
Scharlau, K., 1950, Einfuhrung eines Schwulemasstabes und Abgrenzung von Schwulezonen durch Isohygrothermen, Erdkunde, 4, 188 201.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Sharlau}
\keyword{index}
