% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvol.r
\name{spvol}
\alias{spvol}
\title{spvol}
\usage{
spvol(t, hum_ratio, pair)
}
\arguments{
\item{t}{numeric Air temperature in degC}

\item{hum_ratio}{numeric Humidity ratio}

\item{pair}{numeric Air pressure in (hPa)}
}
\value{
Specific volume.
}
\description{
Gets the specific volume of air/water vapor mixture [m^3/kg].
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{spvol}
