% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssi_index.r
\name{ssi_index}
\alias{ssi_index}
\title{ssi_index}
\usage{
ssi_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Calculate Summer Simmer index developed in Canada.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{SSI}
