% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_indoor.r
\name{steadman_indoor}
\alias{steadman_indoor}
\title{steadman_indoor}
\usage{
steadman_indoor(t, rh)
}
\arguments{
\item{t}{numeric Air temperature in Celsius degrees.}

\item{rh}{numeric  Relative humidity in percentage.}
}
\value{
steadman_indoor
}
\description{
Computes Steadman's apparent temperature for indoor conditions where air movements are negligible.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{apparent}
\keyword{steadman,}
\keyword{temperature,indoor}
