% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_indoor.r
\name{steadman_indoor}
\alias{steadman_indoor}
\title{steadman_indoor}
\usage{
steadman_indoor(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Computes Steadman's apparent temperature for indoor conditions where air movements are negligible.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{steadman_indoor}

