% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_outdoor_shade.r
\name{steadman_outdoor_shade}
\alias{steadman_outdoor_shade}
\title{steadman_outdoor_shade}
\usage{
steadman_outdoor_shade(t, rh, wind)
}
\arguments{
\item{t}{numeric     Air temperature in Celsius degrees.}

\item{rh}{numeric    Relative humidity in percentage.}

\item{wind}{numeric  Wind speed in meters per second.}
}
\value{
Apparent temperature index.
}
\description{
Computes Steadman's apparent temperature under shaded outdoor conditions.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{Steadman,}
\keyword{apparent}
\keyword{outdoor,}
\keyword{shade,}
\keyword{temperature.}
