% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_outdoor_shade.r
\name{steadman_outdoor_shade}
\alias{steadman_outdoor_shade}
\title{steadman_outdoor_shade}
\usage{
steadman_outdoor_shade(t, rh, wind)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Wind speed in meteronseconds ms}
}
\value{
Apparent temperature index.
}
\description{
Computes Steadman's apparent temperature under shade outdoor conditions.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Steadman,outdoor,shade,Apparent}
\keyword{Temeperature.}

