% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadman_outdoor_sun.r
\name{steadman_outdoor_sun}
\alias{steadman_outdoor_sun}
\title{steadman_outdoor_sun}
\usage{
steadman_outdoor_sun(t, rh, wind, rshort, sunelev)
}
\arguments{
\item{t}{numeric       Air temperature in degC.}

\item{rh}{numeric      Relative humidity in percentage.}

\item{wind}{numeric    Wind speed in meters per second.}

\item{rshort}{numeric  Global short radiation in Watt/mq.}

\item{sunelev}{numeric Sun elevation in degrees.}
}
\value{
steadman outdoor index
}
\description{
Computes Steadman's apparent temperature under sunny outdoor conditions.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{Apparent}
\keyword{Sunny}
\keyword{steadman,}
\keyword{temperature,}
