% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_data.r
\name{sun_data}
\alias{sun_data}
\title{sun_data}
\usage{
sun_data(datetime, lat, lon, parameter = "azimuth")
}
\arguments{
\item{datetime}{Datetime in format YYYY-MM-DD HH:MM:SS.}

\item{numeric}{lat Latitude in decimal degrees.}

\item{numeric}{lon Longitude in decimal degrees.}

\item{character}{parameter Six solar parameter are available by name "azimuth","zenith","solarZenith","elevation","declination","JD".}
}
\value{
Parameter indicated in degrees or integer (JD).
}
\description{
Calculate solar parameter for a location and a time.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{parameter}
\keyword{sun}
