% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_data.r
\name{sun_data}
\alias{sun_data}
\title{sun_data}
\usage{
sun_data(datetime, lat, lon, parameter = "azimuth", tz = "GMT")
}
\arguments{
\item{datetime}{character     Datetime as in YYYY-MM-DDTHH:MM:SS considering local time}

\item{lat}{numeric           Latitude in decimal degrees}

\item{lon}{numeric           Longitude in decimal degrees}

\item{parameter}{character    Six solar parameter are available by name "azimuth","zenith","solarZenith","elevation","declination"or "JD"}

\item{tz}{character           time zone.}
}
\value{
value of parameter indicated
}
\description{
Calculate sun parameters for specific location and time.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{parameters}
\keyword{sun}
