% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunposition.r
\name{sunposition}
\alias{sunposition}
\title{sunposition}
\usage{
sunposition(timenow, lat, long)
}
\arguments{
\item{timenow}{time indicat in time format equivalent to "POSIXct" or "POSIXt" object.}

\item{lat}{numeric latitude in decimal degrees}

\item{lon}{numeric  longitude in decimal degrees}
}
\description{
Function to assess sun azimuth and elevation given the time in GMT format.
}
\references{
Mike Birdgeneau 
\url{https://gist.github.com/mikebirdgeneau/e2dbdc8e9c45c1c2b7fc}
}
\author{
Alfonso crisci \email{a.crisci@ibimet.cnr.it} Marco Morabito \email{m.morabito@unifi.it}
}
