% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunposition.r
\name{sunposition}
\alias{sunposition}
\title{sunposition}
\usage{
sunposition(timenow, lat, long)
}
\arguments{
\item{timenow}{datetime    Datetime indicate in time format equivalent to "POSIXct" or "POSIXt" objects.}

\item{lat}{numeric Latitude in decimal degrees.}

\item{long}{numeric Longitude in decimal degrees.}
}
\value{
of  list containing sun position
}
\description{
Function to assess sun azimuth and elevation given the time in GMT format.
}
\references{
Mike Birdgeneau  \url{https://gist.github.com/mikebirdgeneau/e2dbdc8e9c45c1c2b7fc}
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
