% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_comfort.r
\name{ta_comfort}
\alias{ta_comfort}
\title{ta_comfort}
\usage{
ta_comfort(rh, wind, M, iclo, H)
}
\arguments{
\item{rh}{numeric    Relative humidity in percentage.}

\item{wind}{numeric  Wind speed in meter per second.}

\item{M}{numeric     Metabolic rate of the subject in Watt/mq.}

\item{iclo}{numeric Clothing level of subject in clo unit.}

\item{H}{numeric    Timimg in hours.}
}
\value{
comfort temperature in degC
}
\description{
Calculate the temperature of thermal comfort requireed given other parameters and timing in hours.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{ta_comfort}
