% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tglob_cylinder.r
\name{Tglob_cylinder}
\alias{Tglob_cylinder}
\title{Tglob_cylinder}
\usage{
Tglob_cylinder(t, rh, speed, solar, diam, fdir, zenith, albedo)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.a}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{speed Mean Wind speed in meter per seconds [m/s].}

\item{numeric}{pair Air pressure in millibar [hPa].}

\item{numeric}{solar Global solar radionatino in W/mq.}

\item{numeric}{diam Diameter of the sphere in meters. Input example 0.05 = 50 mm.}

\item{numeric}{fdir Fraction on 1 of direct solar respect to global.}

\item{numeric}{zenith Zenith angle was set to 0 degrees}

\item{numeric}{albedo Mean albedo of surround.}
}
\value{

}
\description{
Calculate the globe temperature having diamter. The author of procedure is James C. Liljegren Decision and Information Sciences Division Argonne National Laboratory.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Liljegren J, Carhart R, Lawday P, Tschopp S, Sharp R (2008)  "Modeling Wet Bulb Globe Temperature using Standard Meteorological Measurements" Journal of Occupational and Environmental Hygiene 5: 645-655 \link{www.climatechimp.org}.
}
\keyword{Globometric}
\keyword{sphere.}
\keyword{temperature,}

