% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THI.r
\name{THI}
\alias{THI}
\title{THI}
\usage{
THI(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{
Temperature Humidity Index value.
}
\description{
Given air temperature (Celsius), relative humidity  gives Temperature Humidity Index discomfort. Giles (1990). Is a linearized form.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{Humidity}
\keyword{Index}
\keyword{THI,}
\keyword{Temperature}
