% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THI_kliber.r
\name{THI_kliber}
\alias{THI_kliber}
\title{THI_kliber}
\usage{
THI_kliber(t, rh)
}
\arguments{
\item{t}{numeric  Air temperature in Celsius degrees.}

\item{rh}{numeric Relative humidity in percentage (%).}
}
\value{

}
\description{
Calculate Temperature Humidity Index following Kliber equantions. Adopted by Sardinia ARPA.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{THI_kliber}
