% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THI_kliber.r
\name{THI_kliber}
\alias{THI_kliber}
\title{THI_kliber}
\usage{
THI_kliber(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Calculate Temperature Humidity INdex following Kliber equantions. Adopted by Sardinia ARPA.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{THI_kliber}

