% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thom.r
\name{thom}
\alias{thom}
\title{thom}
\usage{
thom(t, rh, wind = 0.2, pair = 1010)
}
\arguments{
\item{numeric}{t Air temperature in degC.}

\item{numeric}{rh Air Relative humidity in percentage (%).}

\item{numeric}{tg Solar Radiation Radiance in Watt/mq.}

\item{numeric}{pair Air Pressure in hPa. Default is 1010.}
}
\value{
Thom index
}
\description{
Calculate Thom discomfort index by using wet bulb temperature estimation.
}
\references{
Thom, E.C. 1959 The discomfort index. Weatherwise, 12, 57,60
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{discomfort}
\keyword{index}
\keyword{sultryness,}
\keyword{thom,}
