% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thom.r
\name{thom}
\alias{thom}
\title{thom}
\usage{
thom(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degree.}

\item{numeric}{rh Air Relative humidity in percentage.}
}
\value{
Thom index.
}
\description{
Calculate Thom discomfort index by using wet bulb temperature estimation.
}
\references{
Thom, E.C. 1959: The discomfort index. Weatherwise, 12: 57–60
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{discomfort}
\keyword{index}
\keyword{sultryness,}
\keyword{thom,}
