% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twetbulb_j.r
\name{twetbulb_j}
\alias{twetbulb_j}
\title{twetbulb_j}
\usage{
twetbulb_j(t, pair)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{pair}{numeric Air Pressure in hPa. Default is 1010.}
}
\value{
wet bulb temperature
}
\description{
Estimating wet bulb temperature following  empirical equation given by Jeevananda, 1976.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{bulb}
\keyword{temperature}
\keyword{wet}
