% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utci.r
\name{utci}
\alias{utci}
\title{utci}
\usage{
utci(t, rh, wind, tr = NULL)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Air Relative humidity.}

\item{wind}{numeric Windspeed in meters per second.}

\item{tr}{numeric  Mean radiant temperature in degC.}
}
\value{
UTCI index n degC
}
\description{
Calculate Universal Thermal Climate Index ( UTCI) in degC.
}
\references{
Brode P, Jendritzky G, Fiala D and Havenith G, 2011, The Universal Thermal Climate Index UTCI in Operational Use, International Journal of Biometeorology.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{UTCI}
