% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTCI.r
\name{UTCI}
\alias{UTCI}
\title{UTCI}
\usage{
UTCI(ta, rh, wind, tmrt)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in %.}

\item{numeric}{wind Wind speed in meterper second [m/s].}

\item{numeric}{tr Mean radiant temperature in Celsius degrees}
}
\value{
UTCI
}
\description{
Calculate Universal Thermal Climate Index ( UTCI) index.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Bröde P,Jendritzky G,Fiala D and Havenith G, (2011)
The Universal Thermal Climate Index UTCI in Operational Use"  International Journal of Biometeorology
\link{http://www.utci.org/isb/documents/windsor_vers05.pdf}
}
\keyword{UTCI}

