% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTCI.r
\name{UTCI}
\alias{UTCI}
\title{UTCI}
\usage{
UTCI(t, rh, wind, tr)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Wind speed in meter per second.}

\item{numeric}{tr Mean radiant temperature in Celsius degrees}
}
\value{
UTCI
}
\description{
Calculate Universal Thermal Climate Index ( UTCI) index.
}
\references{
Bröde P, Jendritzky G, Fiala D and Havenith G, 2011, The Universal Thermal Climate Index UTCI in Operational Use"  International Journal of Biometeorology.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{UTCI}
