% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utci_class.r
\name{utci_class}
\alias{utci_class}
\title{utci_class}
\usage{
utci_class(t, rh, wind, tmrt)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Wind speed in meter per second.}

\item{numeric}{tmrt Mean radiant temperature in Celsius degrees.}
}
\value{
UTCI index in Celsius degrees.
}
\description{
Calculate ten (10) thermal class of Universal Thermal Climate Index ( UTCI) index.
}
\references{
Brode P,Jendritzky G,Fiala D and Havenith G, 2011,The Universal Thermal Climate Index UTCI in Operational Use".International Journal of Biometeorology.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{UTCI}
