% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utci_class7.r
\name{utci_class7}
\alias{utci_class7}
\title{utci_class7}
\usage{
utci_class7(t, rh, wind, tmrt)
}
\arguments{
\item{t}{numeric  Air temperature in degC}

\item{rh}{numeric  Air Relative humidity in percentage (%)}

\item{wind}{numeric  Wind speed in meters per second}

\item{tmrt}{numeric  Mean radiant temperature in degC.}
}
\value{
UTCI class.
}
\description{
Calculate 7-fold thermal classe by using value of  Universal Thermal Climate Index  UTCI index.
}
\references{
Brode P,Jendritzky G,Fiala D and Havenith G, 2011,The Universal Thermal Climate Index UTCI in Operational Use".International Journal of Biometeorology.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{UTCI}
