% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscosity.r
\name{viscosity}
\alias{viscosity}
\title{viscosity}
\usage{
viscosity(tk)
}
\arguments{
\item{tk}{numeric Air temperature in degK.}
}
\value{
air viscosity
}
\description{
Calculates  the viscosity of air [kg/(m s)] given air temperature in Kelvin degrees.
}
\references{
Oke T R, Boundary Layer Climate,Routledge Atmosphere, Weather and Climate, 2th edition, p.373.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{viscosity}
