% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscosity.r
\name{viscosity}
\alias{viscosity}
\title{viscosity}
\usage{
viscosity(t)
}
\arguments{
\item{numeric}{Ta Air temperature in Kelvin Degrees.}
}
\value{
viscosity
}
\description{
Calculates  the viscosity of air [kg/(m s)] given air temperature in Kelvin degrees.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
Oke T R, Boundary Layer Climate,Routledge Atmosphere, Weather and Climate, 2th edition, p.373.
}
\keyword{viscosity}

