% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpaTorr.r
\name{vpaTorr}
\alias{vpaTorr}
\title{vpaTorr}
\usage{
vpaTorr(T)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}
}
\value{

}
\description{
Calculate water vapor pressure in torr.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{vpaTorr}

