% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpaTorr.r
\name{vpaTorr}
\alias{vpaTorr}
\title{vpaTorr}
\usage{
vpaTorr(t)
}
\arguments{
\item{t}{numeric Air temperature in degC.}
}
\value{
vpa in Torr
}
\description{
Calculate water vapor pressure in torr.
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{vpaTorr}
