% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbdt.r
\name{wbdt}
\alias{wbdt}
\title{wbdt}
\usage{
wbdt(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Given t air temperature (Celsius), rh relative humidity (%)  gives  Wet-bulb dry temperature (WBDT).
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{WBDT}

