% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbdt.r
\name{wbdt}
\alias{wbdt}
\title{wbdt}
\usage{
wbdt(t, rh, pair = 1010)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage (%).}

\item{wind}{numeric Mean Wind speed in meters per second.}
}
\value{
Wet-bulb dry temperature in degC.
}
\description{
Given t air temperature (Celsius), rh relative humidity (percentage)  gives  Wet-bulb dry temperature (WBDT).
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{WBDT}
