% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_indoor.r
\name{wbgt_indoor}
\alias{wbgt_indoor}
\title{wbgt_indoor}
\usage{
wbgt_indoor(t, rh, wind = 0.1, pair = 1010, elev = 0)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{rh}{numeric Relative humidity in percentage .}

\item{wind}{numeric Mean Wind speed in meters per second.}

\item{pair}{numeric Air pressure in millibar or hPa. Default 1010 hPa.}

\item{elev}{numeric Elevation in meters on sea level. Default is 0.}
}
\value{

}
\description{
Calculate web bulb globe index for indoor environments.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{wbgt_indoor}
