% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_indoor.r
\name{wbgt_indoor}
\alias{wbgt_indoor}
\title{wbgt_indoor}
\usage{
wbgt_indoor(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity in %.}
}
\value{

}
\description{
Calculate web bulb globe index for indoor environments.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{wbgt_indoor}

