% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_outdoor.r
\name{wbgt_outdoor}
\alias{wbgt_outdoor}
\title{wbgt_outdoor}
\usage{
wbgt_outdoor(t, rh, wind, solar, press = NULL, topo = NULL)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{wind Air wind wind velocity in meter per second. Default is 0.1.}

\item{numeric}{solar Solar Radiation Radiance in watt on square meter.}

\item{numeric}{press Air Pressure in hPa. Default is 1013.25.}

\item{numeric}{topo Elevation in meters on sea level. Default is 0.}
}
\value{

}
\description{
Calculate wet bulb globe temperature index for outdoor environements by using solar radiation.
Pressure correction follow
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{wbgt_outdoor}
