% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt_sun.r
\name{wbgt_outdoor}
\alias{wbgt_outdoor}
\title{wbgt_sun}
\usage{
wbgt_outdoor(
  t,
  rh,
  wind,
  solar,
  zenith = 0,
  pair = 1010,
  alb_sfc = 0.4,
  fdir = 0.8,
  maxair = 10,
  minair = 2,
  prec = 0.01
)
}
\arguments{
\item{t}{numeric        Air temperature in degC.}

\item{rh}{numeric       Relative humidity in percentage.}

\item{wind}{numeric     Mean Wind speed in meter per second.}

\item{solar}{numeric   Global solar radiation in Watt on mq.}

\item{zenith}{numeric   Zenith angle in decimal degrees.}

\item{pair}{numeric     Air pressure in millibar or hPa. Default 1010 hPa.}

\item{alb_sfc}{numeric  Mean albedo of surroundings. Default is 0.4.}

\item{fdir}{numeric     Ratio of directed solar respect to the diffuse.Default is 0.8.}

\item{maxair}{numeric   Upper bound of search range referred to air temperature in degC. Default is 10.}

\item{minair}{numeric   Lower bound of search range referred to air temperature in degC. Default is 10.}

\item{prec}{numeric      Precision of outcomes.Default is 0.01.}
}
\value{
globe temperature in degC

wbgt for outdoor/sunny conditions in degC
}
\description{
Calculate wet bulb globe temperature index for outdoor environements by using solar radiation.
Pressure correction follow
}
\author{
Istituto per la Bioeconomia CNR Firenze Italy  Alfonso Crisci \email{alfonso.crisci@ibe.cnr.it}
}
\keyword{outdoor}
\keyword{wbgt}
