% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetbulb.r
\name{wetbulb}
\alias{wetbulb}
\title{wetbulb}
\usage{
wetbulb(t, rh, press = NULL)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{press  Air pressure in hPa or millibar.}
}
\value{
Wet bulb temperature in Celsius degrees.
}
\description{
DESCRIPTION
}
\references{
Brice, T., and T. Hall, 2013: The weather calculator. Available online at \url{https://www.weather.gov/epz/wxcalc}.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{wetbulb}
