% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetbulb.r, R/wetbulb_stull.r
\name{wetbulb}
\alias{wetbulb}
\title{wetbulb}
\usage{
wetbulb(t, rh)

wetbulb(t, rh)
}
\arguments{
\item{t}{numeric        Air temperature in degC.}

\item{rh}{numeric       Relative humidity in percentage.}

\item{wind}{numeric     Mean Wind speed in meter per second.}

\item{pair}{numeric     Air pressure in millibar or hPa. Default 1010 hPa.}

\item{solar}{numeric   Global solar radiation in Watt on mq.}

\item{zenith}{numeric   Zenith angle in decimal degrees.}

\item{alb_sfc}{numeric  Mean albedo of surroundings. Default is 0.4.}

\item{fdir}{numeric     Ratio of directed solar respect to the diffuse.Default is 0.8.}

\item{maxair}{numeric   Upper bound of search range referred to air temperature in degC. Default is 10.}

\item{minair}{numeric   Lower bound of search range referred to air temperature in degC. Default is 10.}

\item{prec}{numeric      Precision of outcomes.Default is 0.01.}

\item{t}{numeric      Air temperature in degC.}

\item{rh}{numeric     Relative humidity in percentage ( %).}

\item{press}{numeric  Air pressure in hPa .}
}
\value{
wet-bulb temperature in degC.

Wet bulb temperature in degC
}
\description{
This function calculate the natural wetbulb temperature.

Compute natural wetbulb temperature by using Stull empirical formulation
}
\references{


Stull, R.B., 2011, Meteorology for Scientists and Engineers, 3rd Edition. 938 pages.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}

Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{wetbulb}
