% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetbulb.r
\name{wetbulb}
\alias{wetbulb}
\title{wetbulb}
\usage{
wetbulb(t, rh, pa)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{rh Air Relative humidity in percentage.}

\item{numeric}{pa  Air pressure in hPa or millibar.}

\item{Wet}{bulb temperature in Celsius degrees.}
}
\value{

}
\description{
DESCRIPTION
}
\references{
Brice T, Hall T,2009, Wet-bulb calculator.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{wetbulb}
