% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windchill.r
\name{windchill}
\alias{windchill}
\title{windchill}
\usage{
windchill(t, v)
}
\arguments{
\item{numeric}{t Air temperature in Celsius degrees.}

\item{numeric}{v Windspeed in meter per second.}
}
\value{
Windchill
}
\description{
Calculates the NWS Windchill Temperature (WCT) to estimate the perceived temperature with wind in cold environemnt.Assumes no impact from the sun and having own face model.
}
\references{
New Windchill NOAA calculator \url{http://www.srh.noaa.gov/epz/?n=wxcalc_windchill}
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\keyword{WCT}
\keyword{windchill,}
