% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windchill.r, R/windchill_cla.r
\name{windchill}
\alias{windchill}
\title{windchill}
\usage{
windchill(t, wind)

windchill(t, wind)
}
\arguments{
\item{t}{numeric Air temperature in degC.}

\item{wind}{numeric Windspeed in meters per second.}

\item{t}{numeric    Air temperature in degC.}

\item{wind}{numeric Windspeed in meters per second.}
}
\value{
windchill index

windchill index in degC
}
\description{
Calculates the NWS Windchill Temperature (WCT) to estimate the perceived temperature with wind in cold environemnt.

Calculates the risk class of Windchill index.
}
\references{
Windchill NOAA calculator \url{https://www.weather.gov/epz/wxcalc_windchill}

Windchill Tables \url{https://www.canada.ca/en/environment-climate-change/services/weather-health/wind-chill-cold-weather/wind-chill-index.html}
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}

Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{classes}
\keyword{degC.}
\keyword{in}
\keyword{index}
\keyword{windchill}
