% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winddir.r
\name{winddir}
\alias{winddir}
\title{winddir}
\usage{
winddir(u, v)
}
\arguments{
\item{u}{numeric U zonal component (m/s) .}

\item{v}{numeric V meridian component (m/s).}
}
\description{
Calculate meteorological wind direction from cartesian components u and v.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{Meteorological}
\keyword{Wind}
\keyword{direction}
