% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windspeed.r
\name{windspeed}
\alias{windspeed}
\title{windspeed}
\usage{
windspeed(u, v)
}
\arguments{
\item{u}{numeric U zonal component (m/s),}

\item{v}{numeric V meridian component (m/s),}
}
\value{

}
\description{
Calculate meteorological wind speed.
}
\author{
Istituto per la Bioeconomia Firenze Italy  Alfonso Crisci \email{a.crisci@ibe.cnr.it}
}
\keyword{Scaler}
\keyword{speed}
\keyword{wind}
